@@ -5,8 +5,8 @@ issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+
 version: 0.7.0
 
 environment:
-  sdk: ^3.6.0
-  flutter: ">=3.27.0"
+  sdk: ^3.8.1
+  flutter: ">=3.32.8"
 
 flutter:
   plugin:
@@ -31,7 +31,7 @@ dev_dependencies:
     sdk: flutter
   leak_tracker_flutter_testing: any
   mockito: ^5.4.4
-  pigeon: ^9.1.0
+  pigeon: ^25.3.2
 
 topics:
   - camera
