@@ -1,86 +1,90 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 package io.flutter.plugins.camerax;
 
-import android.app.Activity;
+import android.annotation.SuppressLint;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.res.Configuration;
 import android.view.Display;
+import android.view.OrientationEventListener;
 import android.view.Surface;
-import android.view.WindowManager;
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
 import io.flutter.embedding.engine.systemchannels.PlatformChannel;
 import io.flutter.embedding.engine.systemchannels.PlatformChannel.DeviceOrientation;
+import java.util.Objects;
 
 /**
  * Support class to help to determine the media orientation based on the orientation of the device.
  */
 public class DeviceOrientationManager {
-
-  interface DeviceOrientationChangeCallback {
-    void onChange(DeviceOrientation newOrientation);
-  }
-
   private static final IntentFilter orientationIntentFilter =
       new IntentFilter(Intent.ACTION_CONFIGURATION_CHANGED);
 
-  private final Activity activity;
-  private final boolean isFrontFacing;
-  private final int sensorOrientation;
-  private final DeviceOrientationChangeCallback deviceOrientationChangeCallback;
+  private final DeviceOrientationManagerProxyApi api;
   private PlatformChannel.DeviceOrientation lastOrientation;
   private BroadcastReceiver broadcastReceiver;
 
-  DeviceOrientationManager(
-      @NonNull Activity activity,
-      boolean isFrontFacing,
-      int sensorOrientation,
-      DeviceOrientationChangeCallback callback) {
-    this.activity = activity;
-    this.isFrontFacing = isFrontFacing;
-    this.sensorOrientation = sensorOrientation;
-    this.deviceOrientationChangeCallback = callback;
+  @VisibleForTesting @Nullable protected OrientationEventListener orientationEventListener;
+
+  DeviceOrientationManager(DeviceOrientationManagerProxyApi api) {
+    this.api = api;
+  }
+
+  @NonNull
+  Context getContext() {
+    return api.getPigeonRegistrar().getContext();
   }
 
   /**
-   * Starts listening to the device's sensors or UI for orientation updates.
+   * Starts listening to the device's sensors for device orientation updates.
    *
    * <p>When orientation information is updated, the callback method of the {@link
-   * DeviceOrientationChangeCallback} is called with the new orientation.
-   *
-   * <p>If the device's ACCELEROMETER_ROTATION setting is enabled the {@link
-   * DeviceOrientationManager} will report orientation updates based on the sensor information. If
-   * the ACCELEROMETER_ROTATION is disabled the {@link DeviceOrientationManager} will fallback to
-   * the deliver orientation updates based on the UI orientation.
+   * DeviceOrientationManagerProxyApi} is called with the new orientation.
    */
+  @SuppressLint(
+      "UnprotectedReceiver") // orientationIntentFilter only listens to protected broadcast
   public void start() {
-    if (broadcastReceiver != null) {
-      return;
-    }
-    broadcastReceiver =
-        new BroadcastReceiver() {
-          @Override
-          public void onReceive(Context context, Intent intent) {
-            handleUIOrientationChange();
-          }
-        };
-    activity.registerReceiver(broadcastReceiver, orientationIntentFilter);
-    broadcastReceiver.onReceive(activity, null);
+    stop();
+
+    // Listen for changes in device orientation at the default rate that is suitable for monitoring
+    // typical screen orientation changes.
+    orientationEventListener = createOrientationEventListener();
+    orientationEventListener.enable();
+  }
+
+  @VisibleForTesting
+  @NonNull
+  /**
+   * Creates an {@link OrientationEventListener} that will call the callback method of the {@link
+   * DeviceOrientationManagerProxyApi} whenever it is notified of a new device orientation and this
+   * {@code DeviceOrientationManager} instance determines that the orientation of the device {@link
+   * Configuration} has changed.
+   */
+  protected OrientationEventListener createOrientationEventListener() {
+    return new OrientationEventListener(getContext()) {
+      @Override
+      public void onOrientationChanged(int orientation) {
+        handleUiOrientationChange();
+      }
+    };
   }
 
   /** Stops listening for orientation updates. */
   public void stop() {
-    if (broadcastReceiver == null) {
+    if (orientationEventListener == null) {
       return;
     }
-    activity.unregisterReceiver(broadcastReceiver);
-    broadcastReceiver = null;
+    lastOrientation = null;
+
+    orientationEventListener.disable();
+    orientationEventListener = null;
   }
 
   /**
@@ -90,26 +94,44 @@ public class DeviceOrientationManager {
    * class.
    */
   @VisibleForTesting
-  void handleUIOrientationChange() {
-    PlatformChannel.DeviceOrientation orientation = getUIOrientation();
-    handleOrientationChange(orientation, lastOrientation, deviceOrientationChangeCallback);
+  void handleUiOrientationChange() {
+    PlatformChannel.DeviceOrientation orientation = getUiOrientation();
+    handleOrientationChange(this, orientation, lastOrientation, api);
     lastOrientation = orientation;
   }
 
   /**
-   * Handles orientation changes coming from either the device's sensors or the
-   * OrientationIntentFilter.
+   * Handles orientation changes coming from the device's sensors.
    *
    * <p>This method is visible for testing purposes only and should never be used outside this
    * class.
    */
   @VisibleForTesting
   static void handleOrientationChange(
+      DeviceOrientationManager manager,
       DeviceOrientation newOrientation,
       DeviceOrientation previousOrientation,
-      DeviceOrientationChangeCallback callback) {
+      DeviceOrientationManagerProxyApi api) {
     if (!newOrientation.equals(previousOrientation)) {
-      callback.onChange(newOrientation);
+      api.getPigeonRegistrar()
+          .runOnMainThread(
+              new ProxyApiRegistrar.FlutterMethodRunnable() {
+                @Override
+                public void run() {
+                  api.onDeviceOrientationChanged(
+                      manager,
+                      newOrientation.toString(),
+                      ResultCompat.asCompatCallback(
+                          result -> {
+                            if (result.isFailure()) {
+                              onFailure(
+                                  "DeviceOrientationManager.onDeviceOrientationChanged",
+                                  Objects.requireNonNull(result.exceptionOrNull()));
+                            }
+                            return null;
+                          }));
+                }
+              });
     }
   }
 
@@ -124,9 +146,9 @@ public class DeviceOrientationManager {
   // Configuration.ORIENTATION_SQUARE is deprecated.
   @SuppressWarnings("deprecation")
   @NonNull
-  PlatformChannel.DeviceOrientation getUIOrientation() {
+  PlatformChannel.DeviceOrientation getUiOrientation() {
     final int rotation = getDefaultRotation();
-    final int orientation = activity.getResources().getConfiguration().orientation;
+    final int orientation = getContext().getResources().getConfiguration().orientation;
 
     switch (orientation) {
       case Configuration.ORIENTATION_PORTRAIT:
@@ -171,9 +193,8 @@ public class DeviceOrientationManager {
    *
    * @return An instance of the Android {@link android.view.Display}.
    */
-  @SuppressWarnings("deprecation")
   @VisibleForTesting
   Display getDisplay() {
-    return ((WindowManager) activity.getSystemService(Context.WINDOW_SERVICE)).getDefaultDisplay();
+    return api.getPigeonRegistrar().getDisplay();
   }
 }
