@@ -1,66 +1,36 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 package io.flutter.plugins.camerax;
 
-import static android.hardware.camera2.CameraMetadata.CONTROL_VIDEO_STABILIZATION_MODE_OFF;
-import static android.hardware.camera2.CameraMetadata.CONTROL_VIDEO_STABILIZATION_MODE_ON;
-import static android.hardware.camera2.CameraMetadata.CONTROL_VIDEO_STABILIZATION_MODE_PREVIEW_STABILIZATION;
-import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.when;
 
 import android.hardware.camera2.CameraCharacteristics;
 import android.hardware.camera2.CameraMetadata;
 import androidx.camera.camera2.interop.Camera2CameraInfo;
 import androidx.camera.core.CameraInfo;
-import io.flutter.plugin.common.BinaryMessenger;
-import java.util.List;
-import java.util.Objects;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Rule;
+import java.util.Map;
 import org.junit.Test;
-import org.mockito.Mock;
+import org.junit.runner.RunWith;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
-import org.mockito.junit.MockitoJUnit;
-import org.mockito.junit.MockitoRule;
 import org.mockito.stubbing.Answer;
+import org.robolectric.RobolectricTestRunner;
+import org.robolectric.annotation.Config;
 
+@RunWith(RobolectricTestRunner.class)
 public class Camera2CameraInfoTest {
-  @Rule public MockitoRule mockitoRule = MockitoJUnit.rule();
-
-  @Mock public Camera2CameraInfo mockCamera2CameraInfo;
-
-  InstanceManager testInstanceManager;
-
-  @Before
-  public void setUp() {
-    testInstanceManager = InstanceManager.create(identifier -> {});
-  }
-
-  @After
-  public void tearDown() {
-    testInstanceManager.stopFinalizationListener();
-  }
-
   @Test
-  public void createFrom_createsInstanceFromCameraInfoInstance() {
-    final Camera2CameraInfoHostApiImpl hostApi =
-        new Camera2CameraInfoHostApiImpl(mock(BinaryMessenger.class), testInstanceManager);
-    final long camera2CameraInfoIdentifier = 60;
-    final CameraInfo mockCameraInfo = mock(CameraInfo.class);
-    final long cameraInfoIdentifier = 92;
+  public void from_createsInstanceFromCameraInfoInstance() {
+    final PigeonApiCamera2CameraInfo api =
+        new TestProxyApiRegistrar().getPigeonApiCamera2CameraInfo();
 
-    testInstanceManager.addDartCreatedInstance(mockCameraInfo, cameraInfoIdentifier);
-    testInstanceManager.addDartCreatedInstance(mockCamera2CameraInfo, camera2CameraInfoIdentifier);
+    final CameraInfo mockCameraInfo = mock(CameraInfo.class);
+    final Camera2CameraInfo mockCamera2CameraInfo = mock(Camera2CameraInfo.class);
 
     try (MockedStatic<Camera2CameraInfo> mockedCamera2CameraInfo =
         Mockito.mockStatic(Camera2CameraInfo.class)) {
@@ -68,98 +38,70 @@ public class Camera2CameraInfoTest {
           .when(() -> Camera2CameraInfo.from(mockCameraInfo))
           .thenAnswer((Answer<Camera2CameraInfo>) invocation -> mockCamera2CameraInfo);
 
-      hostApi.createFrom(cameraInfoIdentifier);
-      assertEquals(
-          testInstanceManager.getInstance(camera2CameraInfoIdentifier), mockCamera2CameraInfo);
+      assertEquals(api.from(mockCameraInfo), mockCamera2CameraInfo);
     }
   }
 
   @Test
-  public void getSupportedHardwareLevel_returnsExpectedLevel() {
-    final Camera2CameraInfoHostApiImpl hostApi =
-        new Camera2CameraInfoHostApiImpl(mock(BinaryMessenger.class), testInstanceManager);
-    final long camera2CameraInfoIdentifier = 3;
-    final int expectedHardwareLevel = CameraMetadata.INFO_SUPPORTED_HARDWARE_LEVEL_FULL;
-
-    testInstanceManager.addDartCreatedInstance(mockCamera2CameraInfo, camera2CameraInfoIdentifier);
-    when(mockCamera2CameraInfo.getCameraCharacteristic(
-            CameraCharacteristics.INFO_SUPPORTED_HARDWARE_LEVEL))
-        .thenReturn(expectedHardwareLevel);
-
-    assertEquals(
-        expectedHardwareLevel,
-        hostApi.getSupportedHardwareLevel(camera2CameraInfoIdentifier).intValue());
-  }
-
-  @Test
-  public void getAvailableVideoStabilizationModes_returnsNone() {
-    final Camera2CameraInfoHostApiImpl hostApi =
-        new Camera2CameraInfoHostApiImpl(mock(BinaryMessenger.class), testInstanceManager);
-    final long camera2CameraInfoIdentifier = 3;
-
-    testInstanceManager.addDartCreatedInstance(mockCamera2CameraInfo, camera2CameraInfoIdentifier);
-    when(mockCamera2CameraInfo.getCameraCharacteristic(
-            CameraCharacteristics.CONTROL_AVAILABLE_VIDEO_STABILIZATION_MODES))
-        .thenReturn(new int[] {});
-
-    List<Long> returned = hostApi.getAvailableVideoStabilizationModes(camera2CameraInfoIdentifier);
-    assertEquals(0, returned.size());
-  }
-
-  @Test
-  public void getAvailableVideoStabilizationModes_returnsAll() {
-    // arrange
-    final Camera2CameraInfoHostApiImpl hostApi =
-        new Camera2CameraInfoHostApiImpl(mock(BinaryMessenger.class), testInstanceManager);
-    final long camera2CameraInfoIdentifier = 3;
-
-    int[] expected =
-        new int[] {
-          CONTROL_VIDEO_STABILIZATION_MODE_OFF,
-          CONTROL_VIDEO_STABILIZATION_MODE_ON,
-          CONTROL_VIDEO_STABILIZATION_MODE_PREVIEW_STABILIZATION,
-        };
-
-    testInstanceManager.addDartCreatedInstance(mockCamera2CameraInfo, camera2CameraInfoIdentifier);
-    when(mockCamera2CameraInfo.getCameraCharacteristic(
-            CameraCharacteristics.CONTROL_AVAILABLE_VIDEO_STABILIZATION_MODES))
-        .thenReturn(expected);
-
-    // act
-    int[] returned =
-        hostApi
-            .getAvailableVideoStabilizationModes(camera2CameraInfoIdentifier)
-            .stream()
-            .mapToInt(value -> value.intValue())
-            .toArray();
-
-    // assert
-    assertArrayEquals(expected, returned);
-  }
-
-  @Test
   public void getCameraId_returnsExpectedId() {
-    final Camera2CameraInfoHostApiImpl hostApi =
-        new Camera2CameraInfoHostApiImpl(mock(BinaryMessenger.class), testInstanceManager);
-    final long camera2CameraInfoIdentifier = 13;
-    final String expectedCameraId = "testCameraId";
+    final PigeonApiCamera2CameraInfo api =
+        new TestProxyApiRegistrar().getPigeonApiCamera2CameraInfo();
 
-    testInstanceManager.addDartCreatedInstance(mockCamera2CameraInfo, camera2CameraInfoIdentifier);
-    when(mockCamera2CameraInfo.getCameraId()).thenReturn(expectedCameraId);
+    final Camera2CameraInfo instance = mock(Camera2CameraInfo.class);
+    final String value = "myString";
+    when(instance.getCameraId()).thenReturn(value);
 
-    assertEquals(expectedCameraId, hostApi.getCameraId(camera2CameraInfoIdentifier));
+    assertEquals(value, api.getCameraId(instance));
   }
 
+  @Config(minSdk = 28)
+  @SuppressWarnings("unchecked")
   @Test
-  public void flutterApiCreate_makesCallToCreateInstanceOnDartSide() {
-    final Camera2CameraInfoFlutterApiImpl spyFlutterApi =
-        spy(new Camera2CameraInfoFlutterApiImpl(mock(BinaryMessenger.class), testInstanceManager));
+  public void getCameraCharacteristic_returnsCorrespondingValueOfKeyWhenKeyNotRecognized() {
+    final PigeonApiCamera2CameraInfo api =
+        new TestProxyApiRegistrar().getPigeonApiCamera2CameraInfo();
 
-    spyFlutterApi.create(mockCamera2CameraInfo, reply -> {});
+    final Camera2CameraInfo instance = mock(Camera2CameraInfo.class);
+    final CameraCharacteristics.Key<String> key = CameraCharacteristics.INFO_VERSION;
+    final String value = "version info";
+    when(instance.getCameraCharacteristic(key)).thenReturn(value);
 
-    final long identifier =
-        Objects.requireNonNull(
-            testInstanceManager.getIdentifierForStrongReference(mockCamera2CameraInfo));
-    verify(spyFlutterApi).create(eq(identifier), any());
+    assertEquals(value, api.getCameraCharacteristic(instance, key));
+  }
+
+  @Config(minSdk = 23)
+  @SuppressWarnings("unchecked")
+  @Test
+  public void getCameraCharacteristic_returnsExpectedCameraHardwareLevelWhenRequested() {
+    final PigeonApiCamera2CameraInfo api =
+        new TestProxyApiRegistrar().getPigeonApiCamera2CameraInfo();
+
+    final Camera2CameraInfo instance = mock(Camera2CameraInfo.class);
+    final CameraCharacteristics.Key<Integer> key =
+        CameraCharacteristics.INFO_SUPPORTED_HARDWARE_LEVEL;
+
+    // Test known values.
+    Map<Integer, InfoSupportedHardwareLevel> cameraHardwareLevelsToPigeonConstants =
+        Map.of(
+            CameraMetadata.INFO_SUPPORTED_HARDWARE_LEVEL_3, InfoSupportedHardwareLevel.LEVEL3,
+            CameraMetadata.INFO_SUPPORTED_HARDWARE_LEVEL_EXTERNAL,
+                InfoSupportedHardwareLevel.EXTERNAL,
+            CameraMetadata.INFO_SUPPORTED_HARDWARE_LEVEL_FULL, InfoSupportedHardwareLevel.FULL,
+            CameraMetadata.INFO_SUPPORTED_HARDWARE_LEVEL_LEGACY, InfoSupportedHardwareLevel.LEGACY,
+            CameraMetadata.INFO_SUPPORTED_HARDWARE_LEVEL_LIMITED,
+                InfoSupportedHardwareLevel.LIMITED);
+
+    for (int cameraHardwareLevel : cameraHardwareLevelsToPigeonConstants.keySet()) {
+      when(instance.getCameraCharacteristic(key)).thenReturn(cameraHardwareLevel);
+      assertEquals(
+          cameraHardwareLevelsToPigeonConstants.get(cameraHardwareLevel),
+          api.getCameraCharacteristic(instance, key));
+      reset(instance);
+    }
+
+    // Test unknown value.
+    int testUnknownValue = -1;
+    when(instance.getCameraCharacteristic(key)).thenReturn(testUnknownValue);
+    assertEquals(testUnknownValue, api.getCameraCharacteristic(instance, key));
   }
 }
