@@ -1,4 +1,4 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -19,7 +19,7 @@ public class CameraPermissionsManagerTest {
   public void listener_respondsOnce() {
     final int[] calledCounter = {0};
     CameraRequestPermissionsListener permissionsListener =
-        new CameraRequestPermissionsListener((String code, String desc) -> calledCounter[0]++);
+        new CameraRequestPermissionsListener((CameraPermissionsError error) -> calledCounter[0]++);
 
     permissionsListener.onRequestPermissionsResult(
         9796, null, new int[] {PackageManager.PERMISSION_DENIED});
@@ -39,7 +39,9 @@ public class CameraPermissionsManagerTest {
         9796, null, new int[] {PackageManager.PERMISSION_DENIED});
 
     verify(fakeResultCallback)
-        .onResult("CameraAccessDenied", "Camera access permission was denied.");
+        .onResult(
+            new CameraPermissionsError(
+                "CameraAccessDenied", "Camera access permission was denied."));
   }
 
   @Test
@@ -53,7 +55,9 @@ public class CameraPermissionsManagerTest {
         null,
         new int[] {PackageManager.PERMISSION_GRANTED, PackageManager.PERMISSION_DENIED});
 
-    verify(fakeResultCallback).onResult("AudioAccessDenied", "Audio access permission was denied.");
+    verify(fakeResultCallback)
+        .onResult(
+            new CameraPermissionsError("AudioAccessDenied", "Audio access permission was denied."));
   }
 
   @Test
@@ -68,9 +72,12 @@ public class CameraPermissionsManagerTest {
         new int[] {PackageManager.PERMISSION_GRANTED, PackageManager.PERMISSION_GRANTED});
 
     verify(fakeResultCallback, never())
-        .onResult("CameraAccessDenied", "Camera access permission was denied.");
+        .onResult(
+            new CameraPermissionsError(
+                "CameraAccessDenied", "Camera access permission was denied."));
     verify(fakeResultCallback, never())
-        .onResult("AudioAccessDenied", "Audio access permission was denied.");
+        .onResult(
+            new CameraPermissionsError("AudioAccessDenied", "Audio access permission was denied."));
   }
 
   @Test
@@ -84,6 +91,8 @@ public class CameraPermissionsManagerTest {
     permissionsListener.onRequestPermissionsResult(9796, null, new int[] {});
 
     verify(fakeResultCallback)
-        .onResult("CameraAccessDenied", "Camera access permission was denied.");
+        .onResult(
+            new CameraPermissionsError(
+                "CameraAccessDenied", "Camera access permission was denied."));
   }
 }
