@@ -1,221 +1,237 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 package io.flutter.plugins.camerax;
 
+import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
 import android.content.Context;
 import android.view.Surface;
+import androidx.annotation.NonNull;
 import androidx.camera.core.ImageCapture;
 import androidx.camera.core.ImageCaptureException;
 import androidx.camera.core.resolutionselector.ResolutionSelector;
-import io.flutter.plugin.common.BinaryMessenger;
 import java.io.File;
 import java.io.IOException;
 import java.util.concurrent.Executor;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Rule;
+import kotlin.Result;
+import kotlin.Unit;
+import kotlin.jvm.functions.Function1;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.mockito.MockedStatic;
-import org.mockito.junit.MockitoJUnit;
-import org.mockito.junit.MockitoRule;
 import org.robolectric.RobolectricTestRunner;
 
 @RunWith(RobolectricTestRunner.class)
 public class ImageCaptureTest {
-  @Rule public MockitoRule mockitoRule = MockitoJUnit.rule();
+  @Test
+  public void pigeon_defaultConstructor_createsImageCaptureWithCorrectConfiguration() {
+    final PigeonApiImageCapture api = new TestProxyApiRegistrar().getPigeonApiImageCapture();
 
-  @Mock public ImageCapture mockImageCapture;
-  @Mock public BinaryMessenger mockBinaryMessenger;
-  @Mock public CameraXProxy mockCameraXProxy;
+    final ResolutionSelector mockResolutionSelector = new ResolutionSelector.Builder().build();
+    final long targetResolution = Surface.ROTATION_0;
+    final ImageCapture imageCapture =
+        api.pigeon_defaultConstructor(
+            mockResolutionSelector, targetResolution, CameraXFlashMode.OFF);
 
-  InstanceManager testInstanceManager;
-  private Context context;
-  private MockedStatic<File> mockedStaticFile;
-
-  @Before
-  public void setUp() throws Exception {
-    testInstanceManager = spy(InstanceManager.create(identifier -> {}));
-    context = mock(Context.class);
-    mockedStaticFile = mockStatic(File.class);
-  }
-
-  @After
-  public void tearDown() {
-    testInstanceManager.stopFinalizationListener();
-    mockedStaticFile.close();
+    assertEquals(imageCapture.getResolutionSelector(), mockResolutionSelector);
+    assertEquals(imageCapture.getTargetRotation(), Surface.ROTATION_0);
+    assertEquals(imageCapture.getFlashMode(), ImageCapture.FLASH_MODE_OFF);
   }
 
   @Test
-  public void create_createsImageCaptureWithCorrectConfiguration() {
-    final ImageCaptureHostApiImpl imageCaptureHostApiImpl =
-        new ImageCaptureHostApiImpl(mockBinaryMessenger, testInstanceManager, context);
-    final ImageCapture.Builder mockImageCaptureBuilder = mock(ImageCapture.Builder.class);
-    final Long imageCaptureIdentifier = 74L;
-    final int flashMode = ImageCapture.FLASH_MODE_ON;
-    final ResolutionSelector mockResolutionSelector = mock(ResolutionSelector.class);
-    final long mockResolutionSelectorId = 77;
-    final int targetRotation = Surface.ROTATION_270;
+  public void resolutionSelector_returnsExpectedResolutionSelector() {
+    final PigeonApiImageCapture api = new TestProxyApiRegistrar().getPigeonApiImageCapture();
 
-    imageCaptureHostApiImpl.cameraXProxy = mockCameraXProxy;
-    testInstanceManager.addDartCreatedInstance(mockResolutionSelector, mockResolutionSelectorId);
-    when(mockCameraXProxy.createImageCaptureBuilder()).thenReturn(mockImageCaptureBuilder);
-    when(mockImageCaptureBuilder.build()).thenReturn(mockImageCapture);
+    final ImageCapture instance = mock(ImageCapture.class);
+    final ResolutionSelector value = mock(ResolutionSelector.class);
+    when(instance.getResolutionSelector()).thenReturn(value);
 
-    imageCaptureHostApiImpl.create(
-        imageCaptureIdentifier,
-        Long.valueOf(targetRotation),
-        Long.valueOf(flashMode),
-        mockResolutionSelectorId);
-
-    verify(mockImageCaptureBuilder).setTargetRotation(targetRotation);
-    verify(mockImageCaptureBuilder).setFlashMode(flashMode);
-    verify(mockImageCaptureBuilder).setResolutionSelector(mockResolutionSelector);
-    verify(mockImageCaptureBuilder).build();
-    verify(testInstanceManager).addDartCreatedInstance(mockImageCapture, imageCaptureIdentifier);
+    assertEquals(value, api.resolutionSelector(instance));
   }
 
   @Test
   public void setFlashMode_setsFlashModeOfImageCaptureInstance() {
-    final ImageCaptureHostApiImpl imageCaptureHostApiImpl =
-        new ImageCaptureHostApiImpl(mockBinaryMessenger, testInstanceManager, context);
-    final Long imageCaptureIdentifier = 85L;
-    final Long flashMode = Long.valueOf(ImageCapture.FLASH_MODE_AUTO);
+    final PigeonApiImageCapture api = new TestProxyApiRegistrar().getPigeonApiImageCapture();
 
-    testInstanceManager.addDartCreatedInstance(mockImageCapture, imageCaptureIdentifier);
+    final ImageCapture instance = mock(ImageCapture.class);
+    final CameraXFlashMode flashMode = io.flutter.plugins.camerax.CameraXFlashMode.AUTO;
+    api.setFlashMode(instance, flashMode);
 
-    imageCaptureHostApiImpl.setFlashMode(imageCaptureIdentifier, flashMode);
-
-    verify(mockImageCapture).setFlashMode(flashMode.intValue());
+    verify(instance).setFlashMode(ImageCapture.FLASH_MODE_AUTO);
   }
 
   @Test
   public void
       takePicture_sendsRequestToTakePictureWithExpectedConfigurationWhenTemporaryFileCanBeCreated() {
-    final ImageCaptureHostApiImpl imageCaptureHostApiImpl =
-        spy(new ImageCaptureHostApiImpl(mockBinaryMessenger, testInstanceManager, context));
-    final Long imageCaptureIdentifier = 6L;
+    final ProxyApiRegistrar mockApiRegistrar = mock(ProxyApiRegistrar.class);
+    final Context mockContext = mock(Context.class);
     final File mockOutputDir = mock(File.class);
-    final File mockFile = mock(File.class);
-    final ImageCapture.OutputFileOptions mockOutputFileOptions =
-        mock(ImageCapture.OutputFileOptions.class);
-    final ImageCapture.OnImageSavedCallback mockOnImageSavedCallback =
-        mock(ImageCapture.OnImageSavedCallback.class);
-    @SuppressWarnings("unchecked")
-    final GeneratedCameraXLibrary.Result<String> mockResult =
-        mock(GeneratedCameraXLibrary.Result.class);
+    when(mockContext.getCacheDir()).thenReturn(mockOutputDir);
+    when(mockApiRegistrar.getContext()).thenReturn(mockContext);
 
-    testInstanceManager.addDartCreatedInstance(mockImageCapture, imageCaptureIdentifier);
-    when(context.getCacheDir()).thenReturn(mockOutputDir);
-    imageCaptureHostApiImpl.cameraXProxy = mockCameraXProxy;
-    mockedStaticFile
-        .when(
-            () ->
-                File.createTempFile(
-                    ImageCaptureHostApiImpl.TEMPORARY_FILE_NAME,
-                    ImageCaptureHostApiImpl.JPG_FILE_TYPE,
-                    mockOutputDir))
-        .thenReturn(mockFile);
-    when(mockCameraXProxy.createImageCaptureOutputFileOptions(mockFile))
-        .thenReturn(mockOutputFileOptions);
-    when(imageCaptureHostApiImpl.createOnImageSavedCallback(mockFile, mockResult))
-        .thenReturn(mockOnImageSavedCallback);
+    final String filename = "myFile.jpg";
+    final ImageCaptureProxyApi api =
+        new ImageCaptureProxyApi(mockApiRegistrar) {
+          @Override
+          ImageCapture.OutputFileOptions createImageCaptureOutputFileOptions(@NonNull File file) {
+            return super.createImageCaptureOutputFileOptions(file);
+          }
 
-    imageCaptureHostApiImpl.takePicture(imageCaptureIdentifier, mockResult);
+          @NonNull
+          @Override
+          ImageCapture.OnImageSavedCallback createOnImageSavedCallback(
+              @NonNull File file, @NonNull Function1<? super Result<String>, Unit> callback) {
+            final File mockFile = mock(File.class);
+            when(mockFile.getAbsolutePath()).thenReturn(filename);
+            final ImageCapture.OnImageSavedCallback imageSavedCallback =
+                super.createOnImageSavedCallback(mockFile, callback);
+            imageSavedCallback.onImageSaved(mock(ImageCapture.OutputFileResults.class));
+            return imageSavedCallback;
+          }
+        };
 
-    verify(mockImageCapture)
-        .takePicture(eq(mockOutputFileOptions), any(Executor.class), eq(mockOnImageSavedCallback));
+    final ImageCapture instance = mock(ImageCapture.class);
+
+    try (MockedStatic<File> mockedStaticFile = mockStatic(File.class)) {
+      final File mockFile = mock(File.class);
+      mockedStaticFile
+          .when(
+              () ->
+                  File.createTempFile(
+                      ImageCaptureProxyApi.TEMPORARY_FILE_NAME,
+                      ImageCaptureProxyApi.JPG_FILE_TYPE,
+                      mockOutputDir))
+          .thenReturn(mockFile);
+
+      final String[] result = {null};
+      api.takePicture(
+          instance,
+          ResultCompat.asCompatCallback(
+              reply -> {
+                result[0] = reply.getOrNull();
+                return null;
+              }));
+
+      verify(instance)
+          .takePicture(
+              any(ImageCapture.OutputFileOptions.class),
+              any(Executor.class),
+              any(ImageCapture.OnImageSavedCallback.class));
+      assertEquals(result[0], filename);
+    }
   }
 
   @Test
   public void takePicture_sendsErrorWhenTemporaryFileCannotBeCreated() {
-    final ImageCaptureHostApiImpl imageCaptureHostApiImpl =
-        new ImageCaptureHostApiImpl(mockBinaryMessenger, testInstanceManager, context);
-    final Long imageCaptureIdentifier = 6L;
+    final ProxyApiRegistrar mockApiRegistrar = mock(ProxyApiRegistrar.class);
+    final Context mockContext = mock(Context.class);
     final File mockOutputDir = mock(File.class);
-    final File mockTemporaryCaptureFile = mock(File.class);
-    @SuppressWarnings("unchecked")
-    final GeneratedCameraXLibrary.Result<String> mockResult =
-        mock(GeneratedCameraXLibrary.Result.class);
-    final IOException fileCreationException = new IOException();
+    when(mockContext.getCacheDir()).thenReturn(mockOutputDir);
+    when(mockApiRegistrar.getContext()).thenReturn(mockContext);
 
-    testInstanceManager.addDartCreatedInstance(mockImageCapture, imageCaptureIdentifier);
-    imageCaptureHostApiImpl.cameraXProxy = mockCameraXProxy;
-    when(context.getCacheDir()).thenReturn(mockOutputDir);
-    mockedStaticFile
-        .when(
-            () ->
-                File.createTempFile(
-                    ImageCaptureHostApiImpl.TEMPORARY_FILE_NAME,
-                    ImageCaptureHostApiImpl.JPG_FILE_TYPE,
-                    mockOutputDir))
-        .thenThrow(fileCreationException);
+    final PigeonApiImageCapture api = new ImageCaptureProxyApi(mockApiRegistrar);
 
-    imageCaptureHostApiImpl.takePicture(imageCaptureIdentifier, mockResult);
+    final ImageCapture instance = mock(ImageCapture.class);
 
-    verify(mockResult).error(fileCreationException);
-    verify(mockImageCapture, times(0))
-        .takePicture(
-            any(ImageCapture.OutputFileOptions.class),
-            any(Executor.class),
-            any(ImageCapture.OnImageSavedCallback.class));
+    try (MockedStatic<File> mockedStaticFile = mockStatic(File.class)) {
+      final IOException fileCreationException = new IOException();
+      mockedStaticFile
+          .when(
+              () ->
+                  File.createTempFile(
+                      ImageCaptureProxyApi.TEMPORARY_FILE_NAME,
+                      ImageCaptureProxyApi.JPG_FILE_TYPE,
+                      mockOutputDir))
+          .thenThrow(fileCreationException);
+
+      final Throwable[] result = {null};
+      api.takePicture(
+          instance,
+          ResultCompat.asCompatCallback(
+              reply -> {
+                result[0] = reply.exceptionOrNull();
+                return null;
+              }));
+
+      verifyNoInteractions(instance);
+      assertEquals(result[0], fileCreationException);
+    }
   }
 
   @Test
-  public void takePicture_usesExpectedOnImageSavedCallback() {
-    final ImageCaptureHostApiImpl imageCaptureHostApiImpl =
-        new ImageCaptureHostApiImpl(mockBinaryMessenger, testInstanceManager, context);
-    final SystemServicesFlutterApiImpl mockSystemServicesFlutterApiImpl =
-        mock(SystemServicesFlutterApiImpl.class);
-    final File mockFile = mock(File.class);
-    @SuppressWarnings("unchecked")
-    final GeneratedCameraXLibrary.Result<String> mockResult =
-        mock(GeneratedCameraXLibrary.Result.class);
-    final ImageCapture.OutputFileResults mockOutputFileResults =
-        mock(ImageCapture.OutputFileResults.class);
-    final String mockFileAbsolutePath = "absolute/path/to/captured/image";
-    final ImageCaptureException mockException = mock(ImageCaptureException.class);
+  public void takePicture_onImageSavedCallbackCanSendsError() {
+    final ProxyApiRegistrar mockApiRegistrar = mock(ProxyApiRegistrar.class);
+    final Context mockContext = mock(Context.class);
+    final File mockOutputDir = mock(File.class);
+    when(mockContext.getCacheDir()).thenReturn(mockOutputDir);
+    when(mockApiRegistrar.getContext()).thenReturn(mockContext);
 
-    imageCaptureHostApiImpl.cameraXProxy = mockCameraXProxy;
-    when(mockFile.getAbsolutePath()).thenReturn(mockFileAbsolutePath);
+    final ImageCaptureException captureException = mock(ImageCaptureException.class);
+    final ImageCaptureProxyApi api =
+        new ImageCaptureProxyApi(mockApiRegistrar) {
+          @Override
+          ImageCapture.OutputFileOptions createImageCaptureOutputFileOptions(@NonNull File file) {
+            return super.createImageCaptureOutputFileOptions(file);
+          }
 
-    ImageCapture.OnImageSavedCallback onImageSavedCallback =
-        imageCaptureHostApiImpl.createOnImageSavedCallback(mockFile, mockResult);
+          @NonNull
+          @Override
+          ImageCapture.OnImageSavedCallback createOnImageSavedCallback(
+              @NonNull File file, @NonNull Function1<? super Result<String>, Unit> callback) {
+            final ImageCapture.OnImageSavedCallback imageSavedCallback =
+                super.createOnImageSavedCallback(mock(File.class), callback);
+            imageSavedCallback.onError(captureException);
+            return imageSavedCallback;
+          }
+        };
 
-    // Test success case.
-    onImageSavedCallback.onImageSaved(mockOutputFileResults);
+    final ImageCapture instance = mock(ImageCapture.class);
 
-    verify(mockResult).success(mockFileAbsolutePath);
+    try (MockedStatic<File> mockedStaticFile = mockStatic(File.class)) {
+      final File mockFile = mock(File.class);
+      mockedStaticFile
+          .when(
+              () ->
+                  File.createTempFile(
+                      ImageCaptureProxyApi.TEMPORARY_FILE_NAME,
+                      ImageCaptureProxyApi.JPG_FILE_TYPE,
+                      mockOutputDir))
+          .thenReturn(mockFile);
 
-    // Test error case.
-    onImageSavedCallback.onError(mockException);
+      final Throwable[] result = {null};
+      api.takePicture(
+          instance,
+          ResultCompat.asCompatCallback(
+              reply -> {
+                result[0] = reply.exceptionOrNull();
+                return null;
+              }));
 
-    verify(mockResult).error(mockException);
+      verify(instance)
+          .takePicture(
+              any(ImageCapture.OutputFileOptions.class),
+              any(Executor.class),
+              any(ImageCapture.OnImageSavedCallback.class));
+      assertEquals(result[0], captureException);
+    }
   }
 
   @Test
   public void setTargetRotation_makesCallToSetTargetRotation() {
-    final ImageCaptureHostApiImpl hostApi =
-        new ImageCaptureHostApiImpl(mockBinaryMessenger, testInstanceManager, context);
-    final long instanceIdentifier = 42;
-    final int targetRotation = Surface.ROTATION_90;
+    final PigeonApiImageCapture api = new TestProxyApiRegistrar().getPigeonApiImageCapture();
 
-    testInstanceManager.addDartCreatedInstance(mockImageCapture, instanceIdentifier);
+    final ImageCapture instance = mock(ImageCapture.class);
+    final long rotation = 0;
+    api.setTargetRotation(instance, rotation);
 
-    hostApi.setTargetRotation(instanceIdentifier, Long.valueOf(targetRotation));
-
-    verify(mockImageCapture).setTargetRotation(targetRotation);
+    verify(instance).setTargetRotation((int) rotation);
   }
 }
