@@ -1,135 +1,106 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 package io.flutter.plugins.camerax;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import android.Manifest;
 import android.content.Context;
+import android.content.pm.PackageManager;
 import androidx.camera.video.PendingRecording;
 import androidx.camera.video.Recording;
-import androidx.camera.video.VideoRecordEvent;
-import io.flutter.plugin.common.BinaryMessenger;
-import java.util.Objects;
+import androidx.core.content.ContextCompat;
 import java.util.concurrent.Executor;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnit;
-import org.mockito.junit.MockitoRule;
-import org.robolectric.RobolectricTestRunner;
+import org.mockito.MockedStatic;
+import org.mockito.Mockito;
+import org.mockito.stubbing.Answer;
 
-@RunWith(RobolectricTestRunner.class)
 public class PendingRecordingTest {
-  @Rule public MockitoRule mockitoRule = MockitoJUnit.rule();
+  @Test
+  public void withAudioEnabled_enablesAudioWhenRequestedAndPermissionGranted() {
+    final PigeonApiPendingRecording api =
+        new TestProxyApiRegistrar().getPigeonApiPendingRecording();
+    final PendingRecording instance = mock(PendingRecording.class);
+    final PendingRecording newInstance = mock(PendingRecording.class);
 
-  @Mock public BinaryMessenger mockBinaryMessenger;
-  @Mock public PendingRecording mockPendingRecording;
-  @Mock public Recording mockRecording;
-  @Mock public RecordingFlutterApiImpl mockRecordingFlutterApi;
-  @Mock public Context mockContext;
-  @Mock public SystemServicesFlutterApiImpl mockSystemServicesFlutterApi;
-  @Mock public PendingRecordingFlutterApiImpl mockPendingRecordingFlutterApi;
-  @Mock public VideoRecordEvent.Finalize event;
-  @Mock public Throwable throwable;
+    try (MockedStatic<ContextCompat> mockedContextCompat =
+        Mockito.mockStatic(ContextCompat.class)) {
+      mockedContextCompat
+          .when(
+              () ->
+                  ContextCompat.checkSelfPermission(
+                      any(Context.class), eq(Manifest.permission.RECORD_AUDIO)))
+          .thenAnswer((Answer<Integer>) invocation -> PackageManager.PERMISSION_GRANTED);
 
-  InstanceManager testInstanceManager;
+      when(instance.withAudioEnabled(false)).thenReturn(newInstance);
 
-  @Before
-  public void setUp() {
-    testInstanceManager = spy(InstanceManager.create(identifier -> {}));
-  }
-
-  @After
-  public void tearDown() {
-    testInstanceManager.stopFinalizationListener();
+      assertEquals(api.withAudioEnabled(instance, false), newInstance);
+      verify(instance).withAudioEnabled(false);
+    }
   }
 
   @Test
-  public void testStart() {
-    final Long mockPendingRecordingId = 3L;
-    final Long mockRecordingId = testInstanceManager.addHostCreatedInstance(mockRecording);
-    testInstanceManager.addDartCreatedInstance(mockPendingRecording, mockPendingRecordingId);
+  public void withAudioEnabled_doesNotEnableAudioWhenRequestedAndPermissionNotGranted() {
+    final PigeonApiPendingRecording api =
+        new TestProxyApiRegistrar().getPigeonApiPendingRecording();
+    final PendingRecording instance = mock(PendingRecording.class);
+    final PendingRecording newInstance = mock(PendingRecording.class);
 
-    doReturn(mockRecording).when(mockPendingRecording).start(any(), any());
-    doNothing().when(mockRecordingFlutterApi).create(any(Recording.class), any());
-    PendingRecordingHostApiImpl spy =
-        spy(new PendingRecordingHostApiImpl(mockBinaryMessenger, testInstanceManager, mockContext));
-    doReturn(mock(Executor.class)).when(spy).getExecutor();
-    spy.recordingFlutterApi = mockRecordingFlutterApi;
-    assertEquals(spy.start(mockPendingRecordingId), mockRecordingId);
-    verify(mockRecordingFlutterApi).create(eq(mockRecording), any());
+    try (MockedStatic<ContextCompat> mockedContextCompat =
+        Mockito.mockStatic(ContextCompat.class)) {
+      mockedContextCompat
+          .when(
+              () ->
+                  ContextCompat.checkSelfPermission(
+                      any(Context.class), eq(Manifest.permission.RECORD_AUDIO)))
+          .thenAnswer((Answer<Integer>) invocation -> PackageManager.PERMISSION_DENIED);
 
-    testInstanceManager.remove(mockPendingRecordingId);
-    testInstanceManager.remove(mockRecordingId);
+      when(instance.withAudioEnabled(true)).thenReturn(newInstance);
+
+      assertEquals(api.withAudioEnabled(instance, false), newInstance);
+      verify(instance).withAudioEnabled(true);
+    }
   }
 
   @Test
-  public void testHandleVideoRecordEventSendsError() {
-    PendingRecordingHostApiImpl pendingRecordingHostApi =
-        new PendingRecordingHostApiImpl(mockBinaryMessenger, testInstanceManager, mockContext);
-    pendingRecordingHostApi.systemServicesFlutterApi = mockSystemServicesFlutterApi;
-    pendingRecordingHostApi.pendingRecordingFlutterApi = mockPendingRecordingFlutterApi;
-    final String eventMessage = "example failure message";
+  public void withAudioEnabled_doesNotEnableAudioWhenAudioNotRequested() {
+    final PigeonApiPendingRecording api =
+        new TestProxyApiRegistrar().getPigeonApiPendingRecording();
+    final PendingRecording instance = mock(PendingRecording.class);
+    final PendingRecording newInstance = mock(PendingRecording.class);
 
-    when(event.hasError()).thenReturn(true);
-    when(event.getCause()).thenReturn(throwable);
-    when(throwable.toString()).thenReturn(eventMessage);
-    doNothing().when(mockSystemServicesFlutterApi).sendCameraError(any(), any());
+    when(instance.withAudioEnabled(true)).thenReturn(newInstance);
 
-    pendingRecordingHostApi.handleVideoRecordEvent(event);
-
-    verify(mockPendingRecordingFlutterApi).sendVideoRecordingFinalizedEvent(any());
-    verify(mockSystemServicesFlutterApi).sendCameraError(eq(eventMessage), any());
+    assertEquals(api.withAudioEnabled(instance, true), newInstance);
+    verify(instance).withAudioEnabled(true);
   }
 
   @Test
-  public void handleVideoRecordEvent_SendsVideoRecordingFinalizedEvent() {
-    PendingRecordingHostApiImpl pendingRecordingHostApi =
-        new PendingRecordingHostApiImpl(mockBinaryMessenger, testInstanceManager, mockContext);
-    pendingRecordingHostApi.pendingRecordingFlutterApi = mockPendingRecordingFlutterApi;
+  public void start_callsStartOnInstance() {
+    final PigeonApiPendingRecording api =
+        new TestProxyApiRegistrar().getPigeonApiPendingRecording();
 
-    when(event.hasError()).thenReturn(false);
+    final PendingRecording instance = mock(PendingRecording.class);
+    final VideoRecordEventListener listener = event -> {};
+    final Recording value = mock(Recording.class);
 
-    pendingRecordingHostApi.handleVideoRecordEvent(event);
+    try (MockedStatic<ContextCompat> mockedContextCompat =
+        Mockito.mockStatic(ContextCompat.class)) {
+      mockedContextCompat
+          .when(() -> ContextCompat.getMainExecutor(any()))
+          .thenAnswer((Answer<Executor>) invocation -> mock(Executor.class));
 
-    verify(mockPendingRecordingFlutterApi).sendVideoRecordingFinalizedEvent(any());
-  }
+      when(instance.start(any(Executor.class), any())).thenReturn(value);
 
-  @Test
-  public void handleVideoRecordEvent_SendsVideoRecordingStartedEvent() {
-    PendingRecordingHostApiImpl pendingRecordingHostApi =
-        new PendingRecordingHostApiImpl(mockBinaryMessenger, testInstanceManager, mockContext);
-    pendingRecordingHostApi.pendingRecordingFlutterApi = mockPendingRecordingFlutterApi;
-    VideoRecordEvent.Start mockStartEvent = mock(VideoRecordEvent.Start.class);
-
-    pendingRecordingHostApi.handleVideoRecordEvent(mockStartEvent);
-
-    verify(mockPendingRecordingFlutterApi).sendVideoRecordingStartedEvent(any());
-  }
-
-  @Test
-  public void flutterApiCreateTest() {
-    final PendingRecordingFlutterApiImpl spyPendingRecordingFlutterApi =
-        spy(new PendingRecordingFlutterApiImpl(mockBinaryMessenger, testInstanceManager));
-
-    spyPendingRecordingFlutterApi.create(mockPendingRecording, reply -> {});
-
-    final long identifier =
-        Objects.requireNonNull(
-            testInstanceManager.getIdentifierForStrongReference(mockPendingRecording));
-    verify(spyPendingRecordingFlutterApi).create(eq(identifier), any());
+      assertEquals(value, api.start(instance, listener));
+    }
   }
 }
