@@ -1,134 +1,120 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 package io.flutter.plugins.camerax;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.anyString;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.reset;
-import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
 import android.util.Size;
 import android.view.Surface;
+import androidx.annotation.NonNull;
 import androidx.camera.core.Preview;
+import androidx.camera.core.ResolutionInfo;
 import androidx.camera.core.SurfaceRequest;
 import androidx.camera.core.resolutionselector.ResolutionSelector;
 import androidx.core.util.Consumer;
-import io.flutter.plugin.common.BinaryMessenger;
-import io.flutter.plugins.camerax.GeneratedCameraXLibrary.ResolutionInfo;
-import io.flutter.plugins.camerax.GeneratedCameraXLibrary.SystemServicesFlutterApi.Reply;
 import io.flutter.view.TextureRegistry;
 import java.util.concurrent.Executor;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
-import org.mockito.ArgumentMatchers;
-import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnit;
-import org.mockito.junit.MockitoRule;
 import org.robolectric.RobolectricTestRunner;
 
 @RunWith(RobolectricTestRunner.class)
 public class PreviewTest {
-  @Rule public MockitoRule mockitoRule = MockitoJUnit.rule();
+  @Test
+  public void pigeon_defaultConstructor_createsPreviewWithCorrectConfiguration() {
+    final PigeonApiPreview api = new TestProxyApiRegistrar().getPigeonApiPreview();
 
-  @Mock public Preview mockPreview;
-  @Mock public BinaryMessenger mockBinaryMessenger;
-  @Mock public TextureRegistry mockTextureRegistry;
-  @Mock public CameraXProxy mockCameraXProxy;
+    final ResolutionSelector mockResolutionSelector = new ResolutionSelector.Builder().build();
+    final long targetResolution = Surface.ROTATION_0;
+    final Preview instance =
+        api.pigeon_defaultConstructor(mockResolutionSelector, targetResolution);
 
-  InstanceManager testInstanceManager;
-
-  @Before
-  public void setUp() {
-    testInstanceManager = spy(InstanceManager.create(identifier -> {}));
-  }
-
-  @After
-  public void tearDown() {
-    testInstanceManager.stopFinalizationListener();
+    assertEquals(instance.getResolutionSelector(), mockResolutionSelector);
+    assertEquals(instance.getTargetRotation(), Surface.ROTATION_0);
   }
 
   @Test
-  public void create_createsPreviewWithCorrectConfiguration() {
-    final PreviewHostApiImpl previewHostApi =
-        new PreviewHostApiImpl(mockBinaryMessenger, testInstanceManager, mockTextureRegistry);
-    final Preview.Builder mockPreviewBuilder = mock(Preview.Builder.class);
-    final int targetRotation = 90;
-    final Long previewIdentifier = 3L;
-    final ResolutionSelector mockResolutionSelector = mock(ResolutionSelector.class);
-    final long mockResolutionSelectorId = 90;
+  public void resolutionSelector_returnsExpectedResolutionSelector() {
+    final PigeonApiPreview api = new TestProxyApiRegistrar().getPigeonApiPreview();
 
-    previewHostApi.cameraXProxy = mockCameraXProxy;
-    testInstanceManager.addDartCreatedInstance(mockResolutionSelector, mockResolutionSelectorId);
-    when(mockCameraXProxy.createPreviewBuilder()).thenReturn(mockPreviewBuilder);
-    when(mockPreviewBuilder.build()).thenReturn(mockPreview);
+    final Preview instance = mock(Preview.class);
+    final androidx.camera.core.resolutionselector.ResolutionSelector value =
+        mock(ResolutionSelector.class);
+    when(instance.getResolutionSelector()).thenReturn(value);
 
-    previewHostApi.create(
-        previewIdentifier, Long.valueOf(targetRotation), mockResolutionSelectorId);
-
-    verify(mockPreviewBuilder).setTargetRotation(targetRotation);
-    verify(mockPreviewBuilder).setResolutionSelector(mockResolutionSelector);
-    verify(mockPreviewBuilder).build();
-    verify(testInstanceManager).addDartCreatedInstance(mockPreview, previewIdentifier);
+    assertEquals(value, api.resolutionSelector(instance));
   }
 
   @Test
   public void setSurfaceProvider_createsSurfaceProviderAndReturnsTextureEntryId() {
-    final PreviewHostApiImpl previewHostApi =
-        spy(new PreviewHostApiImpl(mockBinaryMessenger, testInstanceManager, mockTextureRegistry));
+    final TextureRegistry mockTextureRegistry = mock(TextureRegistry.class);
     final TextureRegistry.SurfaceProducer mockSurfaceProducer =
         mock(TextureRegistry.SurfaceProducer.class);
-    final Long previewIdentifier = 5L;
-    final Long surfaceProducerEntryId = 120L;
-
-    previewHostApi.cameraXProxy = mockCameraXProxy;
-    testInstanceManager.addDartCreatedInstance(mockPreview, previewIdentifier);
-
+    final long textureId = 0;
+    when(mockSurfaceProducer.id()).thenReturn(textureId);
     when(mockTextureRegistry.createSurfaceProducer()).thenReturn(mockSurfaceProducer);
-    when(mockSurfaceProducer.id()).thenReturn(surfaceProducerEntryId);
+    final PigeonApiPreview api =
+        new TestProxyApiRegistrar() {
+          @NonNull
+          @Override
+          TextureRegistry getTextureRegistry() {
+            return mockTextureRegistry;
+          }
+        }.getPigeonApiPreview();
 
-    final ArgumentCaptor<Preview.SurfaceProvider> surfaceProviderCaptor =
-        ArgumentCaptor.forClass(Preview.SurfaceProvider.class);
+    final Preview instance = mock(Preview.class);
+    final SystemServicesManager systemServicesManager = mock(SystemServicesManager.class);
 
-    // Test that surface provider was set and the surface texture ID was returned.
-    assertEquals(previewHostApi.setSurfaceProvider(previewIdentifier), surfaceProducerEntryId);
-    verify(mockPreview).setSurfaceProvider(surfaceProviderCaptor.capture());
-    verify(previewHostApi).createSurfaceProvider(mockSurfaceProducer);
+    assertEquals(textureId, api.setSurfaceProvider(instance, systemServicesManager));
+    verify(instance).setSurfaceProvider(any(Preview.SurfaceProvider.class));
   }
 
   @Test
   public void createSurfaceProducer_setsExpectedSurfaceProducerCallback() {
-    final PreviewHostApiImpl previewHostApi =
-        new PreviewHostApiImpl(mockBinaryMessenger, testInstanceManager, mockTextureRegistry);
+    final TextureRegistry mockTextureRegistry = mock(TextureRegistry.class);
     final TextureRegistry.SurfaceProducer mockSurfaceProducer =
         mock(TextureRegistry.SurfaceProducer.class);
+    final long textureId = 0;
+    when(mockSurfaceProducer.id()).thenReturn(textureId);
+    when(mockTextureRegistry.createSurfaceProducer()).thenReturn(mockSurfaceProducer);
+    final PreviewProxyApi api =
+        (PreviewProxyApi)
+            new TestProxyApiRegistrar() {
+              @NonNull
+              @Override
+              TextureRegistry getTextureRegistry() {
+                return mockTextureRegistry;
+              }
+            }.getPigeonApiPreview();
+
+    final SystemServicesManager mockSystemServicesManager = mock(SystemServicesManager.class);
     final SurfaceRequest mockSurfaceRequest = mock(SurfaceRequest.class);
     final ArgumentCaptor<TextureRegistry.SurfaceProducer.Callback> callbackCaptor =
         ArgumentCaptor.forClass(TextureRegistry.SurfaceProducer.Callback.class);
 
     when(mockSurfaceRequest.getResolution()).thenReturn(new Size(5, 6));
-    when(mockSurfaceProducer.getSurface()).thenReturn(mock(Surface.class));
+    when(mockSurfaceProducer.getForcedNewSurface()).thenReturn(mock(Surface.class));
 
-    Preview.SurfaceProvider previewSurfaceProvider =
-        previewHostApi.createSurfaceProvider(mockSurfaceProducer);
+    final Preview.SurfaceProvider previewSurfaceProvider =
+        api.createSurfaceProvider(mockSurfaceProducer, mockSystemServicesManager);
     previewSurfaceProvider.onSurfaceRequested(mockSurfaceRequest);
 
     verify(mockSurfaceProducer).setCallback(callbackCaptor.capture());
 
-    TextureRegistry.SurfaceProducer.Callback callback = callbackCaptor.getValue();
+    final TextureRegistry.SurfaceProducer.Callback callback = callbackCaptor.getValue();
 
-    // Verify callback's onSurfaceDestroyed invalidates SurfaceRequest.
-    simulateSurfaceDestruction(callback);
+    // Verify callback's onSurfaceCleanup invalidates SurfaceRequest.
+    simulateSurfaceCleanup(callback);
     verify(mockSurfaceRequest).invalidate();
 
     reset(mockSurfaceRequest);
@@ -138,35 +124,45 @@ public class PreviewTest {
     verifyNoMoreInteractions(mockSurfaceRequest);
   }
 
+  @SuppressWarnings("unchecked")
   @Test
   public void createSurfaceProvider_createsExpectedPreviewSurfaceProvider() {
-    final PreviewHostApiImpl previewHostApi =
-        new PreviewHostApiImpl(mockBinaryMessenger, testInstanceManager, mockTextureRegistry);
+    final TextureRegistry mockTextureRegistry = mock(TextureRegistry.class);
     final TextureRegistry.SurfaceProducer mockSurfaceProducer =
         mock(TextureRegistry.SurfaceProducer.class);
+    final long textureId = 0;
+    when(mockSurfaceProducer.id()).thenReturn(textureId);
+    when(mockTextureRegistry.createSurfaceProducer()).thenReturn(mockSurfaceProducer);
+    final PreviewProxyApi api =
+        (PreviewProxyApi)
+            new TestProxyApiRegistrar() {
+              @NonNull
+              @Override
+              TextureRegistry getTextureRegistry() {
+                return mockTextureRegistry;
+              }
+            }.getPigeonApiPreview();
+
+    final SystemServicesManager mockSystemServicesManager = mock(SystemServicesManager.class);
+
     final Surface mockSurface = mock(Surface.class);
     final SurfaceRequest mockSurfaceRequest = mock(SurfaceRequest.class);
     final SurfaceRequest.Result mockSurfaceRequestResult = mock(SurfaceRequest.Result.class);
-    final SystemServicesFlutterApiImpl mockSystemServicesFlutterApi =
-        mock(SystemServicesFlutterApiImpl.class);
+
     final int resolutionWidth = 200;
     final int resolutionHeight = 500;
     final Long surfaceProducerEntryId = 120L;
 
-    previewHostApi.cameraXProxy = mockCameraXProxy;
     when(mockSurfaceRequest.getResolution())
         .thenReturn(new Size(resolutionWidth, resolutionHeight));
-    when(mockCameraXProxy.createSystemServicesFlutterApiImpl(mockBinaryMessenger))
-        .thenReturn(mockSystemServicesFlutterApi);
-    when(mockSurfaceProducer.getSurface()).thenReturn(mockSurface);
+    when(mockSurfaceProducer.getForcedNewSurface()).thenReturn(mockSurface);
 
     final ArgumentCaptor<Surface> surfaceCaptor = ArgumentCaptor.forClass(Surface.class);
-    @SuppressWarnings("unchecked")
     final ArgumentCaptor<Consumer<SurfaceRequest.Result>> consumerCaptor =
         ArgumentCaptor.forClass(Consumer.class);
 
-    Preview.SurfaceProvider previewSurfaceProvider =
-        previewHostApi.createSurfaceProvider(mockSurfaceProducer);
+    final Preview.SurfaceProvider previewSurfaceProvider =
+        api.createSurfaceProvider(mockSurfaceProducer, mockSystemServicesManager);
     previewSurfaceProvider.onSurfaceRequested(mockSurfaceRequest);
 
     verify(mockSurfaceProducer).setSize(resolutionWidth, resolutionHeight);
@@ -212,76 +208,85 @@ public class PreviewTest {
         .thenReturn(SurfaceRequest.Result.RESULT_INVALID_SURFACE);
     capturedConsumer.accept(mockSurfaceRequestResult);
     verify(mockSurface).release();
-    verify(mockSystemServicesFlutterApi)
-        .sendCameraError(anyString(), ArgumentMatchers.<Reply<Void>>any());
+    verify(mockSystemServicesManager).onCameraError(anyString());
   }
 
   @Test
-  public void releaseFlutterSurfaceTexture_makesCallToReleaseFlutterSurfaceTexture() {
-    final PreviewHostApiImpl previewHostApi =
-        new PreviewHostApiImpl(mockBinaryMessenger, testInstanceManager, mockTextureRegistry);
+  public void releaseSurfaceProvider_makesCallToReleaseFlutterSurfaceTexture() {
+    final TextureRegistry mockTextureRegistry = mock(TextureRegistry.class);
     final TextureRegistry.SurfaceProducer mockSurfaceProducer =
         mock(TextureRegistry.SurfaceProducer.class);
+    when(mockSurfaceProducer.id()).thenReturn(0L);
+    when(mockTextureRegistry.createSurfaceProducer()).thenReturn(mockSurfaceProducer);
+    final PigeonApiPreview api =
+        new TestProxyApiRegistrar() {
+          @NonNull
+          @Override
+          TextureRegistry getTextureRegistry() {
+            return mockTextureRegistry;
+          }
+        }.getPigeonApiPreview();
 
-    previewHostApi.flutterSurfaceProducer = mockSurfaceProducer;
+    final Preview instance = mock(Preview.class);
+    final SystemServicesManager systemServicesManager = mock(SystemServicesManager.class);
+    api.setSurfaceProvider(instance, systemServicesManager);
+    api.releaseSurfaceProvider(instance);
 
-    previewHostApi.releaseFlutterSurfaceTexture();
     verify(mockSurfaceProducer).release();
   }
 
   @Test
-  public void getResolutionInfo_makesCallToRetrievePreviewResolutionInfo() {
-    final PreviewHostApiImpl previewHostApi =
-        new PreviewHostApiImpl(mockBinaryMessenger, testInstanceManager, mockTextureRegistry);
-    final androidx.camera.core.ResolutionInfo mockResolutionInfo =
-        mock(androidx.camera.core.ResolutionInfo.class);
-    final Long previewIdentifier = 23L;
-    final int resolutionWidth = 500;
-    final int resolutionHeight = 200;
+  public void getResolutionInfo_returnsExpectedResolutionInfo() {
+    final PigeonApiPreview api = new TestProxyApiRegistrar().getPigeonApiPreview();
 
-    testInstanceManager.addDartCreatedInstance(mockPreview, previewIdentifier);
-    when(mockPreview.getResolutionInfo()).thenReturn(mockResolutionInfo);
-    when(mockResolutionInfo.getResolution())
-        .thenReturn(new Size(resolutionWidth, resolutionHeight));
+    final Preview instance = mock(Preview.class);
+    final androidx.camera.core.ResolutionInfo value = mock(ResolutionInfo.class);
+    when(instance.getResolutionInfo()).thenReturn(value);
 
-    ResolutionInfo resolutionInfo = previewHostApi.getResolutionInfo(previewIdentifier);
-    assertEquals(resolutionInfo.getWidth(), Long.valueOf(resolutionWidth));
-    assertEquals(resolutionInfo.getHeight(), Long.valueOf(resolutionHeight));
+    assertEquals(value, api.getResolutionInfo(instance));
   }
 
   @Test
-  public void setTargetRotation_makesCallToSetTargetRotation() {
-    final PreviewHostApiImpl hostApi =
-        new PreviewHostApiImpl(mockBinaryMessenger, testInstanceManager, mockTextureRegistry);
-    final long instanceIdentifier = 52;
-    final int targetRotation = Surface.ROTATION_180;
+  public void setTargetRotation_returnsExpectedTargetRotation() {
+    final PigeonApiPreview api = new TestProxyApiRegistrar().getPigeonApiPreview();
 
-    testInstanceManager.addDartCreatedInstance(mockPreview, instanceIdentifier);
+    final Preview instance = mock(Preview.class);
+    final long rotation = 0;
+    api.setTargetRotation(instance, rotation);
 
-    hostApi.setTargetRotation(instanceIdentifier, Long.valueOf(targetRotation));
-
-    verify(mockPreview).setTargetRotation(targetRotation);
+    verify(instance).setTargetRotation((int) rotation);
   }
 
   @Test
   public void
       surfaceProducerHandlesCropAndRotation_returnsIfSurfaceProducerHandlesCropAndRotation() {
-    final PreviewHostApiImpl hostApi =
-        new PreviewHostApiImpl(mockBinaryMessenger, testInstanceManager, mockTextureRegistry);
+    final TextureRegistry mockTextureRegistry = mock(TextureRegistry.class);
     final TextureRegistry.SurfaceProducer mockSurfaceProducer =
         mock(TextureRegistry.SurfaceProducer.class);
+    when(mockSurfaceProducer.id()).thenReturn(0L);
+    when(mockTextureRegistry.createSurfaceProducer()).thenReturn(mockSurfaceProducer);
+    final PigeonApiPreview api =
+        new TestProxyApiRegistrar() {
+          @NonNull
+          @Override
+          TextureRegistry getTextureRegistry() {
+            return mockTextureRegistry;
+          }
+        }.getPigeonApiPreview();
 
-    hostApi.flutterSurfaceProducer = mockSurfaceProducer;
-    when(mockSurfaceProducer.handlesCropAndRotation()).thenReturn(true);
+    final Preview instance = mock(Preview.class);
+    final SystemServicesManager systemServicesManager = mock(SystemServicesManager.class);
+    api.setSurfaceProvider(instance, systemServicesManager);
+    api.surfaceProducerHandlesCropAndRotation(instance);
 
-    assertEquals(hostApi.surfaceProducerHandlesCropAndRotation(), true);
+    verify(mockSurfaceProducer).handlesCropAndRotation();
   }
 
   // TODO(bparrishMines): Replace with inline calls to onSurfaceCleanup once available on stable;
   // see https://github.com/flutter/flutter/issues/16125. This separate method only exists to scope
   // the suppression.
   @SuppressWarnings({"deprecation", "removal"})
-  void simulateSurfaceDestruction(TextureRegistry.SurfaceProducer.Callback producerLifecycle) {
-    producerLifecycle.onSurfaceDestroyed();
+  void simulateSurfaceCleanup(TextureRegistry.SurfaceProducer.Callback producerLifecycle) {
+    producerLifecycle.onSurfaceCleanup();
   }
 }
