@@ -1,71 +1,59 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 package io.flutter.plugins.camerax;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import android.util.Size;
 import androidx.camera.core.resolutionselector.ResolutionStrategy;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnit;
-import org.mockito.junit.MockitoRule;
+import org.junit.runner.RunWith;
+import org.robolectric.RobolectricTestRunner;
 
+@RunWith(RobolectricTestRunner.class)
 public class ResolutionStrategyTest {
-  @Rule public MockitoRule mockitoRule = MockitoJUnit.rule();
-  @Mock public ResolutionStrategy mockResolutionStrategy;
-  @Mock public ResolutionStrategyHostApiImpl.ResolutionStrategyProxy mockProxy;
+  @Test
+  public void
+      pigeon_defaultConstructor_createsExpectedResolutionStrategyInstanceWhenArgumentsValid() {
+    final PigeonApiResolutionStrategy api =
+        new TestProxyApiRegistrar().getPigeonApiResolutionStrategy();
 
-  InstanceManager instanceManager;
+    final Size size = new Size(1, 2);
+    final ResolutionStrategy resolutionStrategy =
+        api.pigeon_defaultConstructor(size, ResolutionStrategyFallbackRule.CLOSEST_HIGHER);
 
-  @Before
-  public void setUp() {
-    instanceManager = InstanceManager.create(identifier -> {});
-  }
-
-  @After
-  public void tearDown() {
-    instanceManager.stopFinalizationListener();
+    assertEquals(resolutionStrategy.getBoundSize(), size);
+    assertEquals(
+        resolutionStrategy.getFallbackRule(), ResolutionStrategy.FALLBACK_RULE_CLOSEST_HIGHER);
   }
 
   @Test
-  public void hostApiCreate_createsExpectedResolutionStrategyInstanceWhenArgumentsValid() {
-    final GeneratedCameraXLibrary.ResolutionInfo boundSize =
-        new GeneratedCameraXLibrary.ResolutionInfo.Builder().setWidth(50L).setHeight(30L).build();
+  public void getBoundSize_returnsExpectedSize() {
+    final PigeonApiResolutionStrategy api =
+        new TestProxyApiRegistrar().getPigeonApiResolutionStrategy();
 
-    final Long fallbackRule = 0L;
+    final ResolutionStrategy instance = mock(ResolutionStrategy.class);
+    final Size value = mock(Size.class);
+    when(instance.getBoundSize()).thenReturn(value);
 
-    when(mockProxy.create(any(Size.class), eq(fallbackRule))).thenReturn(mockResolutionStrategy);
-
-    final ResolutionStrategyHostApiImpl hostApi =
-        new ResolutionStrategyHostApiImpl(instanceManager, mockProxy);
-
-    final long instanceIdentifier = 0;
-    hostApi.create(instanceIdentifier, boundSize, fallbackRule);
-
-    assertEquals(instanceManager.getInstance(instanceIdentifier), mockResolutionStrategy);
+    assertEquals(value, api.getBoundSize(instance));
   }
 
   @Test
-  public void hostApiCreate_throwsAssertionErrorWhenArgumentsInvalid() {
-    final Long fallbackRule = 8L;
-    final long instanceIdentifier = 0;
+  public void getFallbackRule_returnsExpectedFallbackRule() {
+    final PigeonApiResolutionStrategy api =
+        new TestProxyApiRegistrar().getPigeonApiResolutionStrategy();
 
-    final ResolutionStrategyHostApiImpl hostApi =
-        new ResolutionStrategyHostApiImpl(instanceManager, mockProxy);
+    final ResolutionStrategy instance = mock(ResolutionStrategy.class);
+    ;
+    when(instance.getFallbackRule())
+        .thenReturn(ResolutionStrategy.FALLBACK_RULE_CLOSEST_HIGHER_THEN_LOWER);
 
-    // We expect an exception to be thrown if fallback rule is specified but bound size is not.
-    assertThrows(
-        IllegalArgumentException.class,
-        () -> hostApi.create(instanceIdentifier, null, fallbackRule));
+    assertEquals(
+        ResolutionStrategyFallbackRule.CLOSEST_HIGHER_THEN_LOWER, api.getFallbackRule(instance));
   }
 }
