@@ -1,68 +1,79 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 package io.flutter.plugins.camerax;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThrows;
-import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import android.app.Activity;
 import android.content.Context;
-import io.flutter.plugin.common.BinaryMessenger;
-import io.flutter.plugins.camerax.CameraPermissionsManager.PermissionsRegistry;
-import io.flutter.plugins.camerax.CameraPermissionsManager.ResultCallback;
-import io.flutter.plugins.camerax.GeneratedCameraXLibrary.CameraPermissionsErrorData;
-import io.flutter.plugins.camerax.GeneratedCameraXLibrary.Result;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import java.io.File;
 import java.io.IOException;
-import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
-import org.mockito.Mock;
 import org.mockito.MockedStatic;
-import org.mockito.junit.MockitoJUnit;
-import org.mockito.junit.MockitoRule;
 import org.robolectric.RobolectricTestRunner;
 
 @RunWith(RobolectricTestRunner.class)
 public class SystemServicesTest {
-  @Rule public MockitoRule mockitoRule = MockitoJUnit.rule();
-
-  @Mock public BinaryMessenger mockBinaryMessenger;
-  @Mock public InstanceManager mockInstanceManager;
-  @Mock public Context mockContext;
-
   @Test
   public void requestCameraPermissionsTest() {
-    final SystemServicesHostApiImpl systemServicesHostApi =
-        new SystemServicesHostApiImpl(mockBinaryMessenger, mockInstanceManager, mockContext);
-    final CameraXProxy mockCameraXProxy = mock(CameraXProxy.class);
+    final Activity mockActivity = mock(Activity.class);
+    final CameraPermissionsManager.PermissionsRegistry mockPermissionsRegistry =
+        mock(CameraPermissionsManager.PermissionsRegistry.class);
     final CameraPermissionsManager mockCameraPermissionsManager =
         mock(CameraPermissionsManager.class);
-    final Activity mockActivity = mock(Activity.class);
-    final PermissionsRegistry mockPermissionsRegistry = mock(PermissionsRegistry.class);
-    @SuppressWarnings("unchecked")
-    final Result<CameraPermissionsErrorData> mockResult = mock(Result.class);
+    final TestProxyApiRegistrar proxyApiRegistrar =
+        new TestProxyApiRegistrar() {
+          @NonNull
+          @Override
+          public Context getContext() {
+            return mockActivity;
+          }
+
+          @Nullable
+          @Override
+          CameraPermissionsManager.PermissionsRegistry getPermissionsRegistry() {
+            return mockPermissionsRegistry;
+          }
+
+          @NonNull
+          @Override
+          public CameraPermissionsManager getCameraPermissionsManager() {
+            return mockCameraPermissionsManager;
+          }
+        };
+    final SystemServicesManagerProxyApi api = proxyApiRegistrar.getPigeonApiSystemServicesManager();
+
+    final SystemServicesManager instance =
+        new SystemServicesManagerProxyApi.SystemServicesManagerImpl(
+            proxyApiRegistrar.getPigeonApiSystemServicesManager());
     final Boolean enableAudio = false;
 
-    systemServicesHostApi.cameraXProxy = mockCameraXProxy;
-    systemServicesHostApi.setActivity(mockActivity);
-    systemServicesHostApi.setPermissionsRegistry(mockPermissionsRegistry);
-    when(mockCameraXProxy.createCameraPermissionsManager())
-        .thenReturn(mockCameraPermissionsManager);
+    final CameraPermissionsError[] result = {null};
+    api.requestCameraPermissions(
+        instance,
+        enableAudio,
+        ResultCompat.asCompatCallback(
+            reply -> {
+              result[0] = reply.getOrNull();
+              return null;
+            }));
 
-    final ArgumentCaptor<ResultCallback> resultCallbackCaptor =
-        ArgumentCaptor.forClass(ResultCallback.class);
-
-    systemServicesHostApi.requestCameraPermissions(enableAudio, mockResult);
+    final ArgumentCaptor<CameraPermissionsManager.ResultCallback> resultCallbackCaptor =
+        ArgumentCaptor.forClass(CameraPermissionsManager.ResultCallback.class);
 
     // Test camera permissions are requested.
     verify(mockCameraPermissionsManager)
@@ -72,33 +83,39 @@ public class SystemServicesTest {
             eq(enableAudio),
             resultCallbackCaptor.capture());
 
-    ResultCallback resultCallback = resultCallbackCaptor.getValue();
+    CameraPermissionsManager.ResultCallback resultCallback = resultCallbackCaptor.getValue();
 
     // Test no error data is sent upon permissions request success.
-    resultCallback.onResult(null, null);
-    verify(mockResult).success(null);
+    resultCallback.onResult(null);
+    assertNull(result[0]);
 
     // Test expected error data is sent upon permissions request failure.
     final String testErrorCode = "TestErrorCode";
     final String testErrorDescription = "Test error description.";
 
-    final ArgumentCaptor<CameraPermissionsErrorData> cameraPermissionsErrorDataCaptor =
-        ArgumentCaptor.forClass(CameraPermissionsErrorData.class);
+    final ArgumentCaptor<GeneratedCameraXLibrary.CameraPermissionsErrorData>
+        cameraPermissionsErrorDataCaptor =
+            ArgumentCaptor.forClass(GeneratedCameraXLibrary.CameraPermissionsErrorData.class);
 
-    resultCallback.onResult(testErrorCode, testErrorDescription);
-    verify(mockResult, times(2)).success(cameraPermissionsErrorDataCaptor.capture());
-
-    CameraPermissionsErrorData cameraPermissionsErrorData =
-        cameraPermissionsErrorDataCaptor.getValue();
-    assertEquals(cameraPermissionsErrorData.getErrorCode(), testErrorCode);
-    assertEquals(cameraPermissionsErrorData.getDescription(), testErrorDescription);
+    resultCallback.onResult(new CameraPermissionsError(testErrorCode, testErrorDescription));
+    assertEquals(result[0], new CameraPermissionsError(testErrorCode, testErrorDescription));
   }
 
   @Test
   public void getTempFilePath_returnsCorrectPath() {
-    final SystemServicesHostApiImpl systemServicesHostApi =
-        new SystemServicesHostApiImpl(mockBinaryMessenger, mockInstanceManager, mockContext);
+    final Context mockContext = mock(Context.class);
+    final TestProxyApiRegistrar proxyApiRegistrar =
+        new TestProxyApiRegistrar() {
+          @NonNull
+          @Override
+          public Context getContext() {
+            return mockContext;
+          }
+        };
+    final SystemServicesManagerProxyApi api = proxyApiRegistrar.getPigeonApiSystemServicesManager();
 
+    final SystemServicesManager instance =
+        new SystemServicesManagerProxyApi.SystemServicesManagerImpl(api);
     final String prefix = "prefix";
     final String suffix = ".suffix";
     final MockedStatic<File> mockedStaticFile = mockStatic(File.class);
@@ -108,16 +125,28 @@ public class SystemServicesTest {
     mockedStaticFile
         .when(() -> File.createTempFile(prefix, suffix, mockOutputDir))
         .thenReturn(mockFile);
+
     when(mockFile.toString()).thenReturn(prefix + suffix);
-    assertEquals(systemServicesHostApi.getTempFilePath(prefix, suffix), prefix + suffix);
+    assertEquals(api.getTempFilePath(instance, prefix, suffix), prefix + suffix);
 
     mockedStaticFile.close();
   }
 
   @Test
   public void getTempFilePath_throwsRuntimeExceptionOnIOException() {
-    final SystemServicesHostApiImpl systemServicesHostApi =
-        new SystemServicesHostApiImpl(mockBinaryMessenger, mockInstanceManager, mockContext);
+    final Context mockContext = mock(Context.class);
+    final TestProxyApiRegistrar proxyApiRegistrar =
+        new TestProxyApiRegistrar() {
+          @NonNull
+          @Override
+          public Context getContext() {
+            return mockContext;
+          }
+        };
+    final SystemServicesManagerProxyApi api = proxyApiRegistrar.getPigeonApiSystemServicesManager();
+
+    final SystemServicesManager instance =
+        new SystemServicesManagerProxyApi.SystemServicesManagerImpl(api);
 
     final String prefix = "prefix";
     final String suffix = ".suffix";
@@ -127,10 +156,21 @@ public class SystemServicesTest {
     mockedStaticFile
         .when(() -> File.createTempFile(prefix, suffix, mockOutputDir))
         .thenThrow(IOException.class);
-    assertThrows(
-        GeneratedCameraXLibrary.FlutterError.class,
-        () -> systemServicesHostApi.getTempFilePath(prefix, suffix));
+    assertThrows(RuntimeException.class, () -> api.getTempFilePath(instance, prefix, suffix));
 
     mockedStaticFile.close();
+  }
+
+  @Test
+  public void onCameraError() {
+    final SystemServicesManagerProxyApi mockApi = mock(SystemServicesManagerProxyApi.class);
+    when(mockApi.getPigeonRegistrar()).thenReturn(new TestProxyApiRegistrar());
+
+    final SystemServicesManager instance =
+        new SystemServicesManagerProxyApi.SystemServicesManagerImpl(mockApi);
+    final String errorDescription = "myString";
+    instance.onCameraError(errorDescription);
+
+    verify(mockApi).onCameraError(eq(instance), eq(errorDescription), any());
   }
 }
