@@ -1,4 +1,4 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -57,20 +57,20 @@ class CameraValue {
 
   /// Creates a new camera controller state for an uninitialized controller.
   const CameraValue.uninitialized()
-      : this(
-          isInitialized: false,
-          isRecordingVideo: false,
-          isTakingPicture: false,
-          isStreamingImages: false,
-          isRecordingPaused: false,
-          flashMode: FlashMode.auto,
-          exposureMode: ExposureMode.auto,
-          exposurePointSupported: false,
-          focusMode: FocusMode.auto,
-          focusPointSupported: false,
-          deviceOrientation: DeviceOrientation.portraitUp,
-          isPreviewPaused: false,
-        );
+    : this(
+        isInitialized: false,
+        isRecordingVideo: false,
+        isTakingPicture: false,
+        isStreamingImages: false,
+        isRecordingPaused: false,
+        flashMode: FlashMode.auto,
+        exposureMode: ExposureMode.auto,
+        exposurePointSupported: false,
+        focusMode: FocusMode.auto,
+        focusPointSupported: false,
+        deviceOrientation: DeviceOrientation.portraitUp,
+        isPreviewPaused: false,
+      );
 
   /// True after [CameraController.initialize] has completed successfully.
   final bool isInitialized;
@@ -181,16 +181,19 @@ class CameraValue {
           exposurePointSupported ?? this.exposurePointSupported,
       focusPointSupported: focusPointSupported ?? this.focusPointSupported,
       deviceOrientation: deviceOrientation ?? this.deviceOrientation,
-      lockedCaptureOrientation: lockedCaptureOrientation == null
-          ? this.lockedCaptureOrientation
-          : lockedCaptureOrientation.orNull,
-      recordingOrientation: recordingOrientation == null
-          ? this.recordingOrientation
-          : recordingOrientation.orNull,
+      lockedCaptureOrientation:
+          lockedCaptureOrientation == null
+              ? this.lockedCaptureOrientation
+              : lockedCaptureOrientation.orNull,
+      recordingOrientation:
+          recordingOrientation == null
+              ? this.recordingOrientation
+              : recordingOrientation.orNull,
       isPreviewPaused: isPreviewPaused ?? this.isPreviewPaused,
-      previewPauseOrientation: previewPauseOrientation == null
-          ? this.previewPauseOrientation
-          : previewPauseOrientation.orNull,
+      previewPauseOrientation:
+          previewPauseOrientation == null
+              ? this.previewPauseOrientation
+              : previewPauseOrientation.orNull,
     );
   }
 
@@ -255,7 +258,7 @@ class CameraController extends ValueNotifier<CameraValue> {
   StreamSubscription<CameraImageData>? _imageStreamSubscription;
   FutureOr<bool>? _initCalled;
   StreamSubscription<DeviceOrientationChangedEvent>?
-      _deviceOrientationSubscription;
+  _deviceOrientationSubscription;
 
   /// Checks whether [CameraController.dispose] has completed successfully.
   ///
@@ -284,22 +287,21 @@ class CameraController extends ValueNotifier<CameraValue> {
       _deviceOrientationSubscription = CameraPlatform.instance
           .onDeviceOrientationChanged()
           .listen((DeviceOrientationChangedEvent event) {
-        value = value.copyWith(
-          deviceOrientation: event.orientation,
-        );
-      });
+            value = value.copyWith(deviceOrientation: event.orientation);
+          });
 
       _cameraId = await CameraPlatform.instance.createCameraWithSettings(
         description,
         mediaSettings ?? const MediaSettings(),
       );
 
-      _unawaited(CameraPlatform.instance
-          .onCameraInitialized(_cameraId)
-          .first
-          .then((CameraInitializedEvent event) {
-        initializeCompleter.complete(event);
-      }));
+      _unawaited(
+        CameraPlatform.instance.onCameraInitialized(_cameraId).first.then((
+          CameraInitializedEvent event,
+        ) {
+          initializeCompleter.complete(event);
+        }),
+      );
 
       await CameraPlatform.instance.initializeCamera(
         _cameraId,
@@ -308,19 +310,22 @@ class CameraController extends ValueNotifier<CameraValue> {
 
       value = value.copyWith(
         isInitialized: true,
-        previewSize: await initializeCompleter.future
-            .then((CameraInitializedEvent event) => Size(
-                  event.previewWidth,
-                  event.previewHeight,
-                )),
-        exposureMode: await initializeCompleter.future
-            .then((CameraInitializedEvent event) => event.exposureMode),
-        focusMode: await initializeCompleter.future
-            .then((CameraInitializedEvent event) => event.focusMode),
+        previewSize: await initializeCompleter.future.then(
+          (CameraInitializedEvent event) =>
+              Size(event.previewWidth, event.previewHeight),
+        ),
+        exposureMode: await initializeCompleter.future.then(
+          (CameraInitializedEvent event) => event.exposureMode,
+        ),
+        focusMode: await initializeCompleter.future.then(
+          (CameraInitializedEvent event) => event.focusMode,
+        ),
         exposurePointSupported: await initializeCompleter.future.then(
-            (CameraInitializedEvent event) => event.exposurePointSupported),
-        focusPointSupported: await initializeCompleter.future
-            .then((CameraInitializedEvent event) => event.focusPointSupported),
+          (CameraInitializedEvent event) => event.exposurePointSupported,
+        ),
+        focusPointSupported: await initializeCompleter.future.then(
+          (CameraInitializedEvent event) => event.focusPointSupported,
+        ),
       );
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
@@ -352,9 +357,11 @@ class CameraController extends ValueNotifier<CameraValue> {
     try {
       await CameraPlatform.instance.pausePreview(_cameraId);
       value = value.copyWith(
-          isPreviewPaused: true,
-          previewPauseOrientation: Optional<DeviceOrientation>.of(
-              value.lockedCaptureOrientation ?? value.deviceOrientation));
+        isPreviewPaused: true,
+        previewPauseOrientation: Optional<DeviceOrientation>.of(
+          value.lockedCaptureOrientation ?? value.deviceOrientation,
+        ),
+      );
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
     }
@@ -368,8 +375,9 @@ class CameraController extends ValueNotifier<CameraValue> {
     try {
       await CameraPlatform.instance.resumePreview(_cameraId);
       value = value.copyWith(
-          isPreviewPaused: false,
-          previewPauseOrientation: const Optional<DeviceOrientation>.absent());
+        isPreviewPaused: false,
+        previewPauseOrientation: const Optional<DeviceOrientation>.absent(),
+      );
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
     }
@@ -415,8 +423,10 @@ class CameraController extends ValueNotifier<CameraValue> {
   ///
   // TODO(bmparr): Add settings for resolution and fps.
   Future<void> startImageStream(onLatestImageAvailable onAvailable) async {
-    assert(defaultTargetPlatform == TargetPlatform.android ||
-        defaultTargetPlatform == TargetPlatform.iOS);
+    assert(
+      defaultTargetPlatform == TargetPlatform.android ||
+          defaultTargetPlatform == TargetPlatform.iOS,
+    );
     _throwIfNotInitialized('startImageStream');
     if (value.isRecordingVideo) {
       throw CameraException(
@@ -435,8 +445,8 @@ class CameraController extends ValueNotifier<CameraValue> {
       _imageStreamSubscription = CameraPlatform.instance
           .onStreamedFrameAvailable(_cameraId)
           .listen((CameraImageData imageData) {
-        onAvailable(CameraImage.fromPlatformInterface(imageData));
-      });
+            onAvailable(CameraImage.fromPlatformInterface(imageData));
+          });
       value = value.copyWith(isStreamingImages: true);
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
@@ -451,8 +461,10 @@ class CameraController extends ValueNotifier<CameraValue> {
   /// The `stopImageStream` method is only available on Android and iOS (other
   /// platforms won't be supported in current setup).
   Future<void> stopImageStream() async {
-    assert(defaultTargetPlatform == TargetPlatform.android ||
-        defaultTargetPlatform == TargetPlatform.iOS);
+    assert(
+      defaultTargetPlatform == TargetPlatform.android ||
+          defaultTargetPlatform == TargetPlatform.iOS,
+    );
     _throwIfNotInitialized('stopImageStream');
     if (!value.isStreamingImages) {
       throw CameraException(
@@ -477,8 +489,9 @@ class CameraController extends ValueNotifier<CameraValue> {
   ///
   /// The video is returned as a [XFile] after calling [stopVideoRecording].
   /// Throws a [CameraException] if the capture fails.
-  Future<void> startVideoRecording(
-      {onLatestImageAvailable? onAvailable}) async {
+  Future<void> startVideoRecording({
+    onLatestImageAvailable? onAvailable,
+  }) async {
     _throwIfNotInitialized('startVideoRecording');
     if (value.isRecordingVideo) {
       throw CameraException(
@@ -496,13 +509,16 @@ class CameraController extends ValueNotifier<CameraValue> {
 
     try {
       await CameraPlatform.instance.startVideoCapturing(
-          VideoCaptureOptions(_cameraId, streamCallback: streamCallback));
+        VideoCaptureOptions(_cameraId, streamCallback: streamCallback),
+      );
       value = value.copyWith(
-          isRecordingVideo: true,
-          isRecordingPaused: false,
-          recordingOrientation: Optional<DeviceOrientation>.of(
-              value.lockedCaptureOrientation ?? value.deviceOrientation),
-          isStreamingImages: onAvailable != null);
+        isRecordingVideo: true,
+        isRecordingPaused: false,
+        recordingOrientation: Optional<DeviceOrientation>.of(
+          value.lockedCaptureOrientation ?? value.deviceOrientation,
+        ),
+        isStreamingImages: onAvailable != null,
+      );
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
     }
@@ -525,8 +541,9 @@ class CameraController extends ValueNotifier<CameraValue> {
     }
 
     try {
-      final XFile file =
-          await CameraPlatform.instance.stopVideoRecording(_cameraId);
+      final XFile file = await CameraPlatform.instance.stopVideoRecording(
+        _cameraId,
+      );
       value = value.copyWith(
         isRecordingVideo: false,
         recordingOrientation: const Optional<DeviceOrientation>.absent(),
@@ -647,18 +664,14 @@ class CameraController extends ValueNotifier<CameraValue> {
     if (point != null &&
         (point.dx < 0 || point.dx > 1 || point.dy < 0 || point.dy > 1)) {
       throw ArgumentError(
-          'The values of point should be anywhere between (0,0) and (1,1).');
+        'The values of point should be anywhere between (0,0) and (1,1).',
+      );
     }
 
     try {
       await CameraPlatform.instance.setExposurePoint(
         _cameraId,
-        point == null
-            ? null
-            : Point<double>(
-                point.dx,
-                point.dy,
-              ),
+        point == null ? null : Point<double>(point.dx, point.dy),
       );
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
@@ -711,8 +724,10 @@ class CameraController extends ValueNotifier<CameraValue> {
   Future<double> setExposureOffset(double offset) async {
     _throwIfNotInitialized('setExposureOffset');
     // Check if offset is in range
-    final List<double> range = await Future.wait(
-        <Future<double>>[getMinExposureOffset(), getMaxExposureOffset()]);
+    final List<double> range = await Future.wait(<Future<double>>[
+      getMinExposureOffset(),
+      getMaxExposureOffset(),
+    ]);
     if (offset < range[0] || offset > range[1]) {
       throw CameraException(
         'exposureOffsetOutOfBounds',
@@ -746,10 +761,14 @@ class CameraController extends ValueNotifier<CameraValue> {
   Future<void> lockCaptureOrientation([DeviceOrientation? orientation]) async {
     try {
       await CameraPlatform.instance.lockCaptureOrientation(
-          _cameraId, orientation ?? value.deviceOrientation);
+        _cameraId,
+        orientation ?? value.deviceOrientation,
+      );
       value = value.copyWith(
-          lockedCaptureOrientation: Optional<DeviceOrientation>.of(
-              orientation ?? value.deviceOrientation));
+        lockedCaptureOrientation: Optional<DeviceOrientation>.of(
+          orientation ?? value.deviceOrientation,
+        ),
+      );
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
     }
@@ -770,7 +789,8 @@ class CameraController extends ValueNotifier<CameraValue> {
     try {
       await CameraPlatform.instance.unlockCaptureOrientation(_cameraId);
       value = value.copyWith(
-          lockedCaptureOrientation: const Optional<DeviceOrientation>.absent());
+        lockedCaptureOrientation: const Optional<DeviceOrientation>.absent(),
+      );
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
     }
@@ -784,17 +804,13 @@ class CameraController extends ValueNotifier<CameraValue> {
     if (point != null &&
         (point.dx < 0 || point.dx > 1 || point.dy < 0 || point.dy > 1)) {
       throw ArgumentError(
-          'The values of point should be anywhere between (0,0) and (1,1).');
+        'The values of point should be anywhere between (0,0) and (1,1).',
+      );
     }
     try {
       await CameraPlatform.instance.setFocusPoint(
         _cameraId,
-        point == null
-            ? null
-            : Point<double>(
-                point.dx,
-                point.dy,
-              ),
+        point == null ? null : Point<double>(point.dx, point.dy),
       );
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
