@@ -1,4 +1,4 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -23,23 +23,24 @@ class CameraPreview extends StatelessWidget {
   Widget build(BuildContext context) {
     return controller.value.isInitialized
         ? ValueListenableBuilder<CameraValue>(
-            valueListenable: controller,
-            builder: (BuildContext context, Object? value, Widget? child) {
-              return AspectRatio(
-                aspectRatio: _isLandscape()
-                    ? controller.value.aspectRatio
-                    : (1 / controller.value.aspectRatio),
-                child: Stack(
-                  fit: StackFit.expand,
-                  children: <Widget>[
-                    _wrapInRotatedBox(child: controller.buildPreview()),
-                    child ?? Container(),
-                  ],
-                ),
-              );
-            },
-            child: child,
-          )
+          valueListenable: controller,
+          builder: (BuildContext context, Object? value, Widget? child) {
+            return AspectRatio(
+              aspectRatio:
+                  _isLandscape()
+                      ? controller.value.aspectRatio
+                      : (1 / controller.value.aspectRatio),
+              child: Stack(
+                fit: StackFit.expand,
+                children: <Widget>[
+                  _wrapInRotatedBox(child: controller.buildPreview()),
+                  child ?? Container(),
+                ],
+              ),
+            );
+          },
+          child: child,
+        )
         : Container();
   }
 
@@ -48,16 +49,13 @@ class CameraPreview extends StatelessWidget {
       return child;
     }
 
-    return RotatedBox(
-      quarterTurns: _getQuarterTurns(),
-      child: child,
-    );
+    return RotatedBox(quarterTurns: _getQuarterTurns(), child: child);
   }
 
   bool _isLandscape() {
     return <DeviceOrientation>[
       DeviceOrientation.landscapeLeft,
-      DeviceOrientation.landscapeRight
+      DeviceOrientation.landscapeRight,
     ].contains(_getApplicableOrientation());
   }
 
