@@ -1,4 +1,4 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -131,9 +131,7 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Camera example'),
-      ),
+      appBar: AppBar(title: const Text('Camera example')),
       body: Column(
         children: <Widget>[
           Expanded(
@@ -150,9 +148,7 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
               ),
               child: Padding(
                 padding: const EdgeInsets.all(1.0),
-                child: Center(
-                  child: _cameraPreviewWidget(),
-                ),
+                child: Center(child: _cameraPreviewWidget()),
               ),
             ),
           ),
@@ -161,10 +157,7 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
           Padding(
             padding: const EdgeInsets.all(5.0),
             child: Row(
-              children: <Widget>[
-                _cameraTogglesRowWidget(),
-                _thumbnailWidget(),
-              ],
+              children: <Widget>[_cameraTogglesRowWidget(), _thumbnailWidget()],
             ),
           ),
         ],
@@ -192,15 +185,17 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
         child: CameraPreview(
           controller!,
           child: LayoutBuilder(
-              builder: (BuildContext context, BoxConstraints constraints) {
-            return GestureDetector(
-              behavior: HitTestBehavior.opaque,
-              onScaleStart: _handleScaleStart,
-              onScaleUpdate: _handleScaleUpdate,
-              onTapDown: (TapDownDetails details) =>
-                  onViewFinderTap(details, constraints),
-            );
-          }),
+            builder: (BuildContext context, BoxConstraints constraints) {
+              return GestureDetector(
+                behavior: HitTestBehavior.opaque,
+                onScaleStart: _handleScaleStart,
+                onScaleUpdate: _handleScaleUpdate,
+                onTapDown:
+                    (TapDownDetails details) =>
+                        onViewFinderTap(details, constraints),
+              );
+            },
+          ),
         ),
       );
     }
@@ -216,8 +211,10 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
       return;
     }
 
-    _currentScale = (_baseScale * details.scale)
-        .clamp(_minAvailableZoom, _maxAvailableZoom);
+    _currentScale = (_baseScale * details.scale).clamp(
+      _minAvailableZoom,
+      _maxAvailableZoom,
+    );
 
     await controller!.setZoomLevel(_currentScale);
   }
@@ -238,8 +235,9 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
               SizedBox(
                 width: 64.0,
                 height: 64.0,
-                child: (localVideoController == null)
-                    ? (
+                child:
+                    (localVideoController == null)
+                        ? (
                         // The captured image on the web contains a network-accessible URL
                         // pointing to a location within the browser. It may be displayed
                         // either with Image.network or Image.memory after loading the image
@@ -247,16 +245,18 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
                         kIsWeb
                             ? Image.network(imageFile!.path)
                             : Image.file(File(imageFile!.path)))
-                    : Container(
-                        decoration: BoxDecoration(
-                            border: Border.all(color: Colors.pink)),
-                        child: Center(
-                          child: AspectRatio(
+                        : Container(
+                          decoration: BoxDecoration(
+                            border: Border.all(color: Colors.pink),
+                          ),
+                          child: Center(
+                            child: AspectRatio(
                               aspectRatio:
                                   localVideoController.value.aspectRatio,
-                              child: VideoPlayer(localVideoController)),
+                              child: VideoPlayer(localVideoController),
+                            ),
+                          ),
                         ),
-                      ),
               ),
           ],
         ),
@@ -279,20 +279,19 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
             // The exposure and focus mode are currently not supported on the web.
             ...!kIsWeb
                 ? <Widget>[
-                    IconButton(
-                      icon: const Icon(Icons.exposure),
-                      color: Colors.blue,
-                      onPressed: controller != null
-                          ? onExposureModeButtonPressed
-                          : null,
-                    ),
-                    IconButton(
-                      icon: const Icon(Icons.filter_center_focus),
-                      color: Colors.blue,
-                      onPressed:
-                          controller != null ? onFocusModeButtonPressed : null,
-                    )
-                  ]
+                  IconButton(
+                    icon: const Icon(Icons.exposure),
+                    color: Colors.blue,
+                    onPressed:
+                        controller != null ? onExposureModeButtonPressed : null,
+                  ),
+                  IconButton(
+                    icon: const Icon(Icons.filter_center_focus),
+                    color: Colors.blue,
+                    onPressed:
+                        controller != null ? onFocusModeButtonPressed : null,
+                  ),
+                ]
                 : <Widget>[],
             IconButton(
               icon: Icon(enableAudio ? Icons.volume_up : Icons.volume_mute),
@@ -300,13 +299,16 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
               onPressed: controller != null ? onAudioModeButtonPressed : null,
             ),
             IconButton(
-              icon: Icon(controller?.value.isCaptureOrientationLocked ?? false
-                  ? Icons.screen_lock_rotation
-                  : Icons.screen_rotation),
+              icon: Icon(
+                controller?.value.isCaptureOrientationLocked ?? false
+                    ? Icons.screen_lock_rotation
+                    : Icons.screen_rotation,
+              ),
               color: Colors.blue,
-              onPressed: controller != null
-                  ? onCaptureOrientationLockButtonPressed
-                  : null,
+              onPressed:
+                  controller != null
+                      ? onCaptureOrientationLockButtonPressed
+                      : null,
             ),
           ],
         ),
@@ -326,39 +328,47 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
           children: <Widget>[
             IconButton(
               icon: const Icon(Icons.flash_off),
-              color: controller?.value.flashMode == FlashMode.off
-                  ? Colors.orange
-                  : Colors.blue,
-              onPressed: controller != null
-                  ? () => onSetFlashModeButtonPressed(FlashMode.off)
-                  : null,
+              color:
+                  controller?.value.flashMode == FlashMode.off
+                      ? Colors.orange
+                      : Colors.blue,
+              onPressed:
+                  controller != null
+                      ? () => onSetFlashModeButtonPressed(FlashMode.off)
+                      : null,
             ),
             IconButton(
               icon: const Icon(Icons.flash_auto),
-              color: controller?.value.flashMode == FlashMode.auto
-                  ? Colors.orange
-                  : Colors.blue,
-              onPressed: controller != null
-                  ? () => onSetFlashModeButtonPressed(FlashMode.auto)
-                  : null,
+              color:
+                  controller?.value.flashMode == FlashMode.auto
+                      ? Colors.orange
+                      : Colors.blue,
+              onPressed:
+                  controller != null
+                      ? () => onSetFlashModeButtonPressed(FlashMode.auto)
+                      : null,
             ),
             IconButton(
               icon: const Icon(Icons.flash_on),
-              color: controller?.value.flashMode == FlashMode.always
-                  ? Colors.orange
-                  : Colors.blue,
-              onPressed: controller != null
-                  ? () => onSetFlashModeButtonPressed(FlashMode.always)
-                  : null,
+              color:
+                  controller?.value.flashMode == FlashMode.always
+                      ? Colors.orange
+                      : Colors.blue,
+              onPressed:
+                  controller != null
+                      ? () => onSetFlashModeButtonPressed(FlashMode.always)
+                      : null,
             ),
             IconButton(
               icon: const Icon(Icons.highlight),
-              color: controller?.value.flashMode == FlashMode.torch
-                  ? Colors.orange
-                  : Colors.blue,
-              onPressed: controller != null
-                  ? () => onSetFlashModeButtonPressed(FlashMode.torch)
-                  : null,
+              color:
+                  controller?.value.flashMode == FlashMode.torch
+                      ? Colors.orange
+                      : Colors.blue,
+              onPressed:
+                  controller != null
+                      ? () => onSetFlashModeButtonPressed(FlashMode.torch)
+                      : null,
             ),
           ],
         ),
@@ -368,14 +378,16 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
 
   Widget _exposureModeControlRowWidget() {
     final ButtonStyle styleAuto = TextButton.styleFrom(
-      foregroundColor: controller?.value.exposureMode == ExposureMode.auto
-          ? Colors.orange
-          : Colors.blue,
+      foregroundColor:
+          controller?.value.exposureMode == ExposureMode.auto
+              ? Colors.orange
+              : Colors.blue,
     );
     final ButtonStyle styleLocked = TextButton.styleFrom(
-      foregroundColor: controller?.value.exposureMode == ExposureMode.locked
-          ? Colors.orange
-          : Colors.blue,
+      foregroundColor:
+          controller?.value.exposureMode == ExposureMode.locked
+              ? Colors.orange
+              : Colors.blue,
     );
 
     return SizeTransition(
@@ -385,22 +397,24 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
           color: Colors.grey.shade50,
           child: Column(
             children: <Widget>[
-              const Center(
-                child: Text('Exposure Mode'),
-              ),
+              const Center(child: Text('Exposure Mode')),
               Row(
                 mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                 children: <Widget>[
                   TextButton(
                     style: styleAuto,
-                    onPressed: controller != null
-                        ? () =>
-                            onSetExposureModeButtonPressed(ExposureMode.auto)
-                        : null,
+                    onPressed:
+                        controller != null
+                            ? () => onSetExposureModeButtonPressed(
+                              ExposureMode.auto,
+                            )
+                            : null,
                     onLongPress: () {
                       if (controller != null) {
-                        CameraPlatform.instance
-                            .setExposurePoint(controller!.cameraId, null);
+                        CameraPlatform.instance.setExposurePoint(
+                          controller!.cameraId,
+                          null,
+                        );
                         showInSnackBar('Resetting exposure point');
                       }
                     },
@@ -408,24 +422,25 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
                   ),
                   TextButton(
                     style: styleLocked,
-                    onPressed: controller != null
-                        ? () =>
-                            onSetExposureModeButtonPressed(ExposureMode.locked)
-                        : null,
+                    onPressed:
+                        controller != null
+                            ? () => onSetExposureModeButtonPressed(
+                              ExposureMode.locked,
+                            )
+                            : null,
                     child: const Text('LOCKED'),
                   ),
                   TextButton(
                     style: styleLocked,
-                    onPressed: controller != null
-                        ? () => controller!.setExposureOffset(0.0)
-                        : null,
+                    onPressed:
+                        controller != null
+                            ? () => controller!.setExposureOffset(0.0)
+                            : null,
                     child: const Text('RESET OFFSET'),
                   ),
                 ],
               ),
-              const Center(
-                child: Text('Exposure Offset'),
-              ),
+              const Center(child: Text('Exposure Offset')),
               Row(
                 mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                 children: <Widget>[
@@ -435,10 +450,12 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
                     min: _minAvailableExposureOffset,
                     max: _maxAvailableExposureOffset,
                     label: _currentExposureOffset.toString(),
-                    onChanged: _minAvailableExposureOffset ==
-                            _maxAvailableExposureOffset
-                        ? null
-                        : setExposureOffset,
+                    onChanged: (_) {},
+                    onChangeEnd:
+                        _minAvailableExposureOffset ==
+                                _maxAvailableExposureOffset
+                            ? null
+                            : setExposureOffset,
                   ),
                   Text(_maxAvailableExposureOffset.toString()),
                 ],
@@ -452,14 +469,16 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
 
   Widget _focusModeControlRowWidget() {
     final ButtonStyle styleAuto = TextButton.styleFrom(
-      foregroundColor: controller?.value.focusMode == FocusMode.auto
-          ? Colors.orange
-          : Colors.blue,
+      foregroundColor:
+          controller?.value.focusMode == FocusMode.auto
+              ? Colors.orange
+              : Colors.blue,
     );
     final ButtonStyle styleLocked = TextButton.styleFrom(
-      foregroundColor: controller?.value.focusMode == FocusMode.locked
-          ? Colors.orange
-          : Colors.blue,
+      foregroundColor:
+          controller?.value.focusMode == FocusMode.locked
+              ? Colors.orange
+              : Colors.blue,
     );
 
     return SizeTransition(
@@ -469,21 +488,22 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
           color: Colors.grey.shade50,
           child: Column(
             children: <Widget>[
-              const Center(
-                child: Text('Focus Mode'),
-              ),
+              const Center(child: Text('Focus Mode')),
               Row(
                 mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                 children: <Widget>[
                   TextButton(
                     style: styleAuto,
-                    onPressed: controller != null
-                        ? () => onSetFocusModeButtonPressed(FocusMode.auto)
-                        : null,
+                    onPressed:
+                        controller != null
+                            ? () => onSetFocusModeButtonPressed(FocusMode.auto)
+                            : null,
                     onLongPress: () {
                       if (controller != null) {
-                        CameraPlatform.instance
-                            .setFocusPoint(controller!.cameraId, null);
+                        CameraPlatform.instance.setFocusPoint(
+                          controller!.cameraId,
+                          null,
+                        );
                       }
                       showInSnackBar('Resetting focus point');
                     },
@@ -491,9 +511,11 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
                   ),
                   TextButton(
                     style: styleLocked,
-                    onPressed: controller != null
-                        ? () => onSetFocusModeButtonPressed(FocusMode.locked)
-                        : null,
+                    onPressed:
+                        controller != null
+                            ? () =>
+                                onSetFocusModeButtonPressed(FocusMode.locked)
+                            : null,
                     child: const Text('LOCKED'),
                   ),
                 ],
@@ -515,11 +537,12 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
         IconButton(
           icon: const Icon(Icons.camera_alt),
           color: Colors.blue,
-          onPressed: cameraController != null &&
-                  cameraController.value.isInitialized &&
-                  !cameraController.value.isRecordingVideo
-              ? onTakePictureButtonPressed
-              : null,
+          onPressed:
+              cameraController != null &&
+                      cameraController.value.isInitialized &&
+                      !cameraController.value.isRecordingVideo
+                  ? onTakePictureButtonPressed
+                  : null,
         ),
         IconButton(
           icon: const Icon(Icons.videocam),
@@ -528,10 +551,11 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
               cameraController == null ? null : onVideoRecordButtonPressed,
         ),
         IconButton(
-          icon: cameraController != null &&
-                  cameraController.value.isRecordingPaused
-              ? const Icon(Icons.play_arrow)
-              : const Icon(Icons.pause),
+          icon:
+              cameraController != null &&
+                      cameraController.value.isRecordingPaused
+                  ? const Icon(Icons.play_arrow)
+                  : const Icon(Icons.pause),
           color: Colors.blue,
           onPressed: () {
             if (cameraController == null) {
@@ -603,8 +627,9 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
   String timestamp() => DateTime.now().millisecondsSinceEpoch.toString();
 
   void showInSnackBar(String message) {
-    ScaffoldMessenger.of(context)
-        .showSnackBar(SnackBar(content: Text(message)));
+    ScaffoldMessenger.of(
+      context,
+    ).showSnackBar(SnackBar(content: Text(message)));
   }
 
   void onViewFinderTap(TapDownDetails details, BoxConstraints constraints) {
@@ -636,11 +661,12 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
 
     final CameraController cameraController = CameraController(
       cameraDescription,
-      mediaSettings: const MediaSettings(
+      mediaSettings: MediaSettings(
         resolutionPreset: ResolutionPreset.low,
         fps: 15,
         videoBitrate: 200000,
         audioBitrate: 32000,
+        enableAudio: enableAudio,
       ),
       imageFormatGroup: ImageFormatGroup.jpeg,
     );
@@ -654,7 +680,8 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
       }
       if (cameraController.value.hasError) {
         showInSnackBar(
-            'Camera error ${cameraController.value.errorDescription}');
+          'Camera error ${cameraController.value.errorDescription}',
+        );
       }
     });
 
@@ -664,19 +691,20 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
         // The exposure mode is currently not supported on the web.
         ...!kIsWeb
             ? <Future<Object?>>[
-                cameraController.getMinExposureOffset().then(
-                    (double value) => _minAvailableExposureOffset = value),
-                cameraController
-                    .getMaxExposureOffset()
-                    .then((double value) => _maxAvailableExposureOffset = value)
-              ]
+              cameraController.getMinExposureOffset().then(
+                (double value) => _minAvailableExposureOffset = value,
+              ),
+              cameraController.getMaxExposureOffset().then(
+                (double value) => _maxAvailableExposureOffset = value,
+              ),
+            ]
             : <Future<Object?>>[],
-        cameraController
-            .getMaxZoomLevel()
-            .then((double value) => _maxAvailableZoom = value),
-        cameraController
-            .getMinZoomLevel()
-            .then((double value) => _minAvailableZoom = value),
+        cameraController.getMaxZoomLevel().then(
+          (double value) => _maxAvailableZoom = value,
+        ),
+        cameraController.getMinZoomLevel().then(
+          (double value) => _minAvailableZoom = value,
+        ),
       ]);
     } on CameraException catch (e) {
       switch (e.code) {
@@ -768,7 +796,8 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
         } else {
           await cameraController.lockCaptureOrientation();
           showInSnackBar(
-              'Capture orientation locked to ${cameraController.value.lockedCaptureOrientation.toString().split('.').last}');
+            'Capture orientation locked to ${cameraController.value.lockedCaptureOrientation.toString().split('.').last}',
+          );
         }
       }
     } on CameraException catch (e) {
@@ -987,9 +1016,10 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
       return;
     }
 
-    final VideoPlayerController vController = kIsWeb
-        ? VideoPlayerController.networkUrl(Uri.parse(videoFile!.path))
-        : VideoPlayerController.file(File(videoFile!.path));
+    final VideoPlayerController vController =
+        kIsWeb
+            ? VideoPlayerController.networkUrl(Uri.parse(videoFile!.path))
+            : VideoPlayerController.file(File(videoFile!.path));
 
     videoPlayerListener = () {
       if (videoController != null) {
@@ -1047,9 +1077,7 @@ class CameraApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(
-      home: CameraExampleHome(),
-    );
+    return const MaterialApp(home: CameraExampleHome());
   }
 }
 
