@@ -1,370 +1,405 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:ui' show Size;
+import 'package:flutter/services.dart';
 
-import 'package:flutter/services.dart' show DeviceOrientation;
+// ignore_for_file: non_constant_identifier_names
 
-import 'analyzer.dart';
-import 'aspect_ratio_strategy.dart';
-import 'camera2_camera_control.dart';
-import 'camera2_camera_info.dart';
-import 'camera_control.dart';
-import 'camera_info.dart';
-import 'camera_selector.dart';
-import 'camera_state.dart';
-import 'camerax_library.g.dart';
-import 'capture_request_options.dart';
-import 'device_orientation_manager.dart';
-import 'fallback_strategy.dart';
-import 'focus_metering_action.dart';
-import 'image_analysis.dart';
-import 'image_capture.dart';
-import 'image_proxy.dart';
-import 'metering_point.dart';
-import 'observer.dart';
-import 'preview.dart';
-import 'process_camera_provider.dart';
-import 'quality_selector.dart';
-import 'recorder.dart';
-import 'resolution_filter.dart';
-import 'resolution_selector.dart';
-import 'resolution_strategy.dart';
-import 'system_services.dart';
-import 'video_capture.dart';
+import 'camerax_library.dart';
 
-/// Handles `JavaObject` creation and calling their methods that require
-/// testing.
+/// Handles constructing objects and calling static methods for the Android
+/// Interactive Media Ads native library.
 ///
-/// By default, each function will create `JavaObject`s attached to an
-/// `InstanceManager` and call through to the appropriate method.
+/// This class provides dependency injection for the implementations of the
+/// platform interface classes. Improving the ease of unit testing and/or
+/// overriding the underlying Android classes.
+///
+/// By default each function calls the default constructor of the class it
+/// intends to return.
 class CameraXProxy {
-  /// Constructs a [CameraXProxy].
+  /// Constructs an [CameraXProxy].
   CameraXProxy({
-    this.getProcessCameraProvider = _getProcessCameraProvider,
-    this.createCameraSelector = _createAttachedCameraSelector,
-    this.createPreview = _createAttachedPreview,
-    this.createImageCapture = _createAttachedImageCapture,
-    this.createRecorder = _createAttachedRecorder,
-    this.createVideoCapture = _createAttachedVideoCapture,
-    this.createImageAnalysis = _createAttachedImageAnalysis,
-    this.createAnalyzer = _createAttachedAnalyzer,
-    this.createCameraStateObserver = _createAttachedCameraStateObserver,
-    this.createResolutionStrategy = _createAttachedResolutionStrategy,
-    this.createResolutionSelector = _createAttachedResolutionSelector,
-    this.createFallbackStrategy = _createAttachedFallbackStrategy,
-    this.createQualitySelector = _createAttachedQualitySelector,
-    this.requestCameraPermissions = _requestCameraPermissions,
-    this.startListeningForDeviceOrientationChange =
-        _startListeningForDeviceOrientationChange,
-    this.setPreviewSurfaceProvider = _setPreviewSurfaceProvider,
-    this.getDefaultDisplayRotation = _getDefaultDisplayRotation,
-    this.getCamera2CameraControl = _getCamera2CameraControl,
-    this.createCaptureRequestOptions = _createAttachedCaptureRequestOptions,
-    this.createMeteringPoint = _createAttachedMeteringPoint,
-    this.createFocusMeteringAction = _createAttachedFocusMeteringAction,
-    this.createAspectRatioStrategy = _createAttachedAspectRatioStrategy,
-    this.createResolutionFilterWithOnePreferredSize =
-        _createAttachedResolutionFilterWithOnePreferredSize,
-    this.getCamera2CameraInfo = _getCamera2CameraInfo,
-    this.getUiOrientation = _getUiOrientation,
-    this.getSensorOrientation = _getSensorOrientation,
-    this.previewSurfaceProducerHandlesCropAndRotation =
-        _previewSurfaceProducerHandlesCropAndRotation,
+    this.setUpGenericsProxy = setUpGenerics,
+    this.newCameraSize = CameraSize.new,
+    this.newCameraIntegerRange = CameraIntegerRange.new,
+    this.newObserver = Observer.new,
+    this.newCameraSelector = CameraSelector.new,
+    this.newSystemServicesManager = SystemServicesManager.new,
+    this.newDeviceOrientationManager = DeviceOrientationManager.new,
+    this.newPreview = Preview.new,
+    this.withOutputVideoCapture = VideoCapture.withOutput,
+    this.newRecorder = Recorder.new,
+    this.newVideoRecordEventListener = VideoRecordEventListener.new,
+    this.newImageCapture = ImageCapture.new,
+    this.newResolutionStrategy = ResolutionStrategy.new,
+    this.newResolutionSelector = ResolutionSelector.new,
+    this.newAspectRatioStrategy = AspectRatioStrategy.new,
+    this.newImageAnalysis = ImageAnalysis.new,
+    this.newAnalyzer = Analyzer.new,
+    this.fromQualitySelector = QualitySelector.from,
+    this.fromOrderedListQualitySelector = QualitySelector.fromOrderedList,
+    this.higherQualityOrLowerThanFallbackStrategy =
+        FallbackStrategy.higherQualityOrLowerThan,
+    this.higherQualityThanFallbackStrategy = FallbackStrategy.higherQualityThan,
+    this.lowerQualityOrHigherThanFallbackStrategy =
+        FallbackStrategy.lowerQualityOrHigherThan,
+    this.lowerQualityThanFallbackStrategy = FallbackStrategy.lowerQualityThan,
+    this.newFocusMeteringActionBuilder = FocusMeteringActionBuilder.new,
+    this.withModeFocusMeteringActionBuilder =
+        FocusMeteringActionBuilder.withMode,
+    this.newCaptureRequestOptions = CaptureRequestOptions.new,
+    this.fromCamera2CameraControl = Camera2CameraControl.from,
+    this.createWithOnePreferredSizeResolutionFilter =
+        ResolutionFilter.createWithOnePreferredSize,
+    this.fromCamera2CameraInfo = Camera2CameraInfo.from,
+    this.newDisplayOrientedMeteringPointFactory =
+        DisplayOrientedMeteringPointFactory.new,
+    this.getInstanceProcessCameraProvider = ProcessCameraProvider.getInstance,
+    this.getResolutionQualitySelector = QualitySelector.getResolution,
+    this.defaultBackCameraCameraSelector = _defaultBackCameraCameraSelector,
+    this.defaultFrontCameraCameraSelector = _defaultFrontCameraCameraSelector,
+    this.highestAvailableStrategyResolutionStrategy =
+        _highestAvailableStrategyResolutionStrategy,
+    this.ratio_16_9FallbackAutoStrategyAspectRatioStrategy =
+        _ratio_16_9FallbackAutoStrategyAspectRatioStrategy,
+    this.ratio_4_3FallbackAutoStrategyAspectRatioStrategy =
+        _ratio_4_3FallbackAutoStrategyAspectRatioStrategy,
+    this.controlAELockCaptureRequest = _controlAELockCaptureRequest,
+    this.infoSupportedHardwareLevelCameraCharacteristics =
+        _infoSupportedHardwareLevelCameraCharacteristics,
+    this.sensorOrientationCameraCharacteristics =
+        _sensorOrientationCameraCharacteristics,
   });
 
-  /// Returns a [ProcessCameraProvider] instance.
-  Future<ProcessCameraProvider> Function() getProcessCameraProvider;
+  /// Handles adding support for generic classes.
+  final void Function({
+    BinaryMessenger? pigeonBinaryMessenger,
+    PigeonInstanceManager? pigeonInstanceManager,
+  })
+  setUpGenericsProxy;
 
-  /// Returns a [CameraSelector] based on the specified camera lens direction.
-  CameraSelector Function(int cameraSelectorLensDirection) createCameraSelector;
+  /// Constructs [CameraSize].
+  final CameraSize Function({
+    required int width,
+    required int height,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newCameraSize;
 
-  /// Returns a [Preview] configured with the specified target rotation and
-  /// specified [ResolutionSelector].
-  Preview Function(
-    ResolutionSelector? resolutionSelector,
+  /// Constructs [CameraIntegerRange].
+  final CameraIntegerRange Function({
+    required int lower,
+    required int upper,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newCameraIntegerRange;
+
+  /// Constructs [Observer].
+  final Observer<T> Function<T>({
+    required void Function(Observer<T>, T) onChanged,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newObserver;
+
+  /// Constructs [CameraSelector].
+  final CameraSelector Function({
+    LensFacing? requireLensFacing,
+    CameraInfo? cameraInfoForFilter,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newCameraSelector;
+
+  /// Constructs [SystemServicesManager].
+  final SystemServicesManager Function({
+    required void Function(SystemServicesManager, String) onCameraError,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newSystemServicesManager;
+
+  /// Constructs [DeviceOrientationManager].
+  final DeviceOrientationManager Function({
+    required void Function(DeviceOrientationManager, String)
+    onDeviceOrientationChanged,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newDeviceOrientationManager;
+
+  /// Constructs [Preview].
+  final Preview Function({
     int? targetRotation,
-  ) createPreview;
+    ResolutionSelector? resolutionSelector,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newPreview;
 
-  /// Returns an [ImageCapture] configured with specified flash mode and
-  /// the specified [ResolutionSelector].
-  ImageCapture Function(
-          ResolutionSelector? resolutionSelector, int? targetRotation)
-      createImageCapture;
+  /// Constructs [VideoCapture].
+  final VideoCapture Function({
+    required VideoOutput videoOutput,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  withOutputVideoCapture;
 
-  /// Returns a [Recorder] for use in video capture configured with the
-  /// specified [QualitySelector].
-  Recorder Function(QualitySelector? qualitySelector) createRecorder;
+  /// Constructs [Recorder].
+  final Recorder Function({
+    int? aspectRatio,
+    int? targetVideoEncodingBitRate,
+    QualitySelector? qualitySelector,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newRecorder;
 
-  /// Returns a [VideoCapture] associated with the provided [Recorder].
-  Future<VideoCapture> Function(Recorder recorder) createVideoCapture;
+  /// Constructs [VideoRecordEventListener].
+  final VideoRecordEventListener Function({
+    required void Function(VideoRecordEventListener, VideoRecordEvent) onEvent,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newVideoRecordEventListener;
 
-  /// Returns an [ImageAnalysis] configured with the specified
-  /// [ResolutionSelector].
-  ImageAnalysis Function(
-          ResolutionSelector? resolutionSelector, int? targetRotation)
-      createImageAnalysis;
+  /// Constructs [ImageCapture].
+  final ImageCapture Function({
+    int? targetRotation,
+    CameraXFlashMode? flashMode,
+    ResolutionSelector? resolutionSelector,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newImageCapture;
 
-  /// Returns an [Analyzer] configured with the specified callback for
-  /// analyzing [ImageProxy]s.
-  Analyzer Function(Future<void> Function(ImageProxy imageProxy) analyze)
-      createAnalyzer;
+  /// Constructs [ResolutionStrategy].
+  final ResolutionStrategy Function({
+    required CameraSize boundSize,
+    required ResolutionStrategyFallbackRule fallbackRule,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newResolutionStrategy;
 
-  /// Returns an [Observer] of the [CameraState] with the specified callback
-  /// for handling changes in that state.
-  Observer<CameraState> Function(void Function(Object stateAsObject) onChanged)
-      createCameraStateObserver;
+  /// Constructs [ResolutionSelector].
+  final ResolutionSelector Function({
+    AspectRatioStrategy? aspectRatioStrategy,
+    ResolutionStrategy? resolutionStrategy,
+    ResolutionFilter? resolutionFilter,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newResolutionSelector;
 
-  /// Returns a [ResolutionStrategy] configured with the specified bounds for
-  /// choosing a resolution and a fallback rule if achieving a resolution within
-  /// those bounds is not possible.
-  ///
-  /// [highestAvailable] is used to specify whether or not the highest available
-  /// [ResolutionStrategy] should be returned.
-  ResolutionStrategy Function(
-      {bool highestAvailable,
-      Size? boundSize,
-      int? fallbackRule}) createResolutionStrategy;
+  /// Constructs [AspectRatioStrategy].
+  final AspectRatioStrategy Function({
+    required AspectRatio preferredAspectRatio,
+    required AspectRatioStrategyFallbackRule fallbackRule,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newAspectRatioStrategy;
 
-  /// Returns a [ResolutionSelector] configured with the specified
-  /// [ResolutionStrategy], [ResolutionFilter], and [AspectRatioStrategy].
-  ResolutionSelector Function(
-      ResolutionStrategy resolutionStrategy,
-      ResolutionFilter? resolutionFilter,
-      AspectRatioStrategy? aspectRatioStrategy) createResolutionSelector;
+  /// Constructs [ImageAnalysis].
+  final ImageAnalysis Function({
+    int? targetRotation,
+    int? outputImageFormat,
+    ResolutionSelector? resolutionSelector,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newImageAnalysis;
 
-  /// Returns a [FallbackStrategy] configured with the specified [VideoQuality]
-  /// and [VideoResolutionFallbackRule].
-  FallbackStrategy Function(
-          {required VideoQuality quality,
-          required VideoResolutionFallbackRule fallbackRule})
-      createFallbackStrategy;
+  /// Constructs [Analyzer].
+  final Analyzer Function({
+    required void Function(Analyzer, ImageProxy) analyze,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newAnalyzer;
 
-  /// Returns a [QualitySelector] configured with the specified [VideoQuality]
-  /// and [FallbackStrategy].
-  QualitySelector Function(
-      {required VideoQuality videoQuality,
-      required FallbackStrategy fallbackStrategy}) createQualitySelector;
+  /// Constructs [QualitySelector].
+  final QualitySelector Function({
+    required VideoQuality quality,
+    FallbackStrategy? fallbackStrategy,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  fromQualitySelector;
 
-  /// Requests camera permissions.
-  Future<void> Function(bool enableAudio) requestCameraPermissions;
+  /// Constructs [QualitySelector].
+  final QualitySelector Function({
+    required List<VideoQuality> qualities,
+    FallbackStrategy? fallbackStrategy,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  fromOrderedListQualitySelector;
 
-  /// Subscribes the plugin as a listener to changes in device orientation.
-  void Function(bool cameraIsFrontFacing, int sensorOrientation)
-      startListeningForDeviceOrientationChange;
+  /// Constructs [FallbackStrategy].
+  final FallbackStrategy Function({
+    required VideoQuality quality,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  higherQualityOrLowerThanFallbackStrategy;
 
-  /// Sets the surface provider of the specified [Preview] instance and returns
-  /// the ID corresponding to the surface it will provide.
-  Future<int> Function(Preview preview) setPreviewSurfaceProvider;
+  /// Constructs [FallbackStrategy].
+  final FallbackStrategy Function({
+    required VideoQuality quality,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  higherQualityThanFallbackStrategy;
 
-  /// Returns default rotation for [UseCase]s in terms of one of the [Surface]
-  /// rotation constants.
-  Future<int> Function() getDefaultDisplayRotation;
+  /// Constructs [FallbackStrategy].
+  final FallbackStrategy Function({
+    required VideoQuality quality,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  lowerQualityOrHigherThanFallbackStrategy;
 
-  /// Gets [Camera2CameraControl] instance from [cameraControl].
-  Camera2CameraControl Function(CameraControl cameraControl)
-      getCamera2CameraControl;
+  /// Constructs [FallbackStrategy].
+  final FallbackStrategy Function({
+    required VideoQuality quality,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  lowerQualityThanFallbackStrategy;
 
-  /// Creates a [CaptureRequestOptions] with specified options.
-  CaptureRequestOptions Function(
-          List<(CaptureRequestKeySupportedType, Object?)> options)
-      createCaptureRequestOptions;
+  /// Constructs [FocusMeteringActionBuilder].
+  final FocusMeteringActionBuilder Function({
+    required MeteringPoint point,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newFocusMeteringActionBuilder;
 
-  /// Returns a [MeteringPoint] with the specified coordinates based on
-  /// [cameraInfo].
-  MeteringPoint Function(
-          double x, double y, double? size, CameraInfo cameraInfo)
-      createMeteringPoint;
+  /// Constructs [FocusMeteringActionBuilder].
+  FocusMeteringActionBuilder Function({
+    required MeteringPoint point,
+    required MeteringMode mode,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  withModeFocusMeteringActionBuilder;
 
-  /// Returns a [FocusMeteringAction] based on the specified metering points
-  /// and their modes.
-  FocusMeteringAction Function(List<(MeteringPoint, int?)> meteringPointInfos,
-      bool? disableAutoCancel) createFocusMeteringAction;
+  /// Constructs [CaptureRequestOptions].
+  CaptureRequestOptions Function({
+    required Map<CaptureRequestKey, Object?> options,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newCaptureRequestOptions;
 
-  /// Creates an [AspectRatioStrategy] with specified aspect ratio and fallback
-  /// rule.
-  AspectRatioStrategy Function(int aspectRatio, int fallbackRule)
-      createAspectRatioStrategy;
+  /// Constructs [Camera2CameraControl].
+  Camera2CameraControl Function({
+    required CameraControl cameraControl,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  fromCamera2CameraControl;
 
-  /// Creates a [ResolutionFilter] that prioritizes specified resolution.
-  ResolutionFilter Function(Size preferredResolution)
-      createResolutionFilterWithOnePreferredSize;
+  /// Constructs [ResolutionFilter].
+  final ResolutionFilter Function({
+    required CameraSize preferredSize,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  createWithOnePreferredSizeResolutionFilter;
 
-  /// Gets [Camera2CameraInfo] instance from [cameraInfo].
-  Future<Camera2CameraInfo> Function(CameraInfo cameraInfo)
-      getCamera2CameraInfo;
+  /// Constructs [Camera2CameraInfo].
+  final Camera2CameraInfo Function({
+    required CameraInfo cameraInfo,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  fromCamera2CameraInfo;
 
-  /// Gets current UI orientation based on device orientation and rotation.
-  Future<DeviceOrientation> Function() getUiOrientation;
+  /// Constructs [DisplayOrientedMeteringPointFactory].
+  DisplayOrientedMeteringPointFactory Function({
+    required CameraInfo cameraInfo,
+    required double width,
+    required double height,
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  newDisplayOrientedMeteringPointFactory;
 
-  /// Gets camera sensor orientation from [camera2CameraInfo].
-  Future<int> Function(Camera2CameraInfo camera2CameraInfo)
-      getSensorOrientation;
+  /// Calls to [ProcessCameraProvider.getInstance].
+  final Future<ProcessCameraProvider> Function({
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  getInstanceProcessCameraProvider;
 
-  /// Returns whether or not the preview's surface producer handles correctly
-  /// rotating the camera preview automatically.
-  Future<bool> Function(Preview preview)
-      previewSurfaceProducerHandlesCropAndRotation;
+  /// Calls to [QualitySelector.getResolution].
+  final Future<CameraSize?> Function(
+    CameraInfo,
+    VideoQuality, {
+    BinaryMessenger? pigeon_binaryMessenger,
+    PigeonInstanceManager? pigeon_instanceManager,
+  })
+  getResolutionQualitySelector;
 
-  static Future<ProcessCameraProvider> _getProcessCameraProvider() {
-    return ProcessCameraProvider.getInstance();
-  }
+  /// Calls to [CameraSelector.defaultBackCamera].
+  final CameraSelector Function() defaultBackCameraCameraSelector;
 
-  static CameraSelector _createAttachedCameraSelector(
-      int cameraSelectorLensDirection) {
-    switch (cameraSelectorLensDirection) {
-      case CameraSelector.lensFacingFront:
-        return CameraSelector.getDefaultFrontCamera();
-      case CameraSelector.lensFacingBack:
-        return CameraSelector.getDefaultBackCamera();
-      default:
-        return CameraSelector(lensFacing: cameraSelectorLensDirection);
-    }
-  }
+  /// Calls to [CameraSelector.defaultFrontCamera].
+  final CameraSelector Function() defaultFrontCameraCameraSelector;
 
-  static Preview _createAttachedPreview(
-      ResolutionSelector? resolutionSelector, int? targetRotation) {
-    return Preview(
-        initialTargetRotation: targetRotation,
-        resolutionSelector: resolutionSelector);
-  }
+  /// Calls to [ResolutionStrategy.highestAvailableStrategy].
+  final ResolutionStrategy Function()
+  highestAvailableStrategyResolutionStrategy;
 
-  static ImageCapture _createAttachedImageCapture(
-      ResolutionSelector? resolutionSelector, int? targetRotation) {
-    return ImageCapture(
-        resolutionSelector: resolutionSelector,
-        initialTargetRotation: targetRotation);
-  }
+  /// Calls to [AspectRatioStrategy.ratio_16_9FallbackAutoStrategy].
+  final AspectRatioStrategy Function()
+  ratio_16_9FallbackAutoStrategyAspectRatioStrategy;
 
-  static Recorder _createAttachedRecorder(QualitySelector? qualitySelector) {
-    return Recorder(qualitySelector: qualitySelector);
-  }
+  /// Calls to [AspectRatioStrategy.ratio_4_3FallbackAutoStrategy].
+  final AspectRatioStrategy Function()
+  ratio_4_3FallbackAutoStrategyAspectRatioStrategy;
 
-  static Future<VideoCapture> _createAttachedVideoCapture(
-      Recorder recorder) async {
-    return VideoCapture.withOutput(recorder);
-  }
+  /// Calls to [CaptureRequest.controlAELock].
+  CaptureRequestKey Function() controlAELockCaptureRequest;
 
-  static ImageAnalysis _createAttachedImageAnalysis(
-      ResolutionSelector? resolutionSelector, int? targetRotation) {
-    return ImageAnalysis(
-        resolutionSelector: resolutionSelector,
-        initialTargetRotation: targetRotation);
-  }
+  /// Calls to [CameraCharacteristics.infoSupportedHardwareLevel].
+  final CameraCharacteristicsKey Function()
+  infoSupportedHardwareLevelCameraCharacteristics;
 
-  static Analyzer _createAttachedAnalyzer(
-      Future<void> Function(ImageProxy imageProxy) analyze) {
-    return Analyzer(analyze: analyze);
-  }
+  /// Calls to [CameraCharacteristics.sensorOrientation].
+  final CameraCharacteristicsKey Function()
+  sensorOrientationCameraCharacteristics;
 
-  static Observer<CameraState> _createAttachedCameraStateObserver(
-      void Function(Object stateAsObject) onChanged) {
-    return Observer<CameraState>(onChanged: onChanged);
-  }
+  static CameraSelector _defaultBackCameraCameraSelector() =>
+      CameraSelector.defaultBackCamera;
 
-  static ResolutionStrategy _createAttachedResolutionStrategy(
-      {bool highestAvailable = false, Size? boundSize, int? fallbackRule}) {
-    if (highestAvailable) {
-      return ResolutionStrategy.highestAvailableStrategy();
-    }
+  static CameraSelector _defaultFrontCameraCameraSelector() =>
+      CameraSelector.defaultFrontCamera;
 
-    return ResolutionStrategy(
-        boundSize: boundSize!, fallbackRule: fallbackRule);
-  }
+  static ResolutionStrategy _highestAvailableStrategyResolutionStrategy() =>
+      ResolutionStrategy.highestAvailableStrategy;
 
-  static ResolutionSelector _createAttachedResolutionSelector(
-      ResolutionStrategy resolutionStrategy,
-      ResolutionFilter? resolutionFilter,
-      AspectRatioStrategy? aspectRatioStrategy) {
-    return ResolutionSelector(
-        resolutionStrategy: resolutionStrategy,
-        resolutionFilter: resolutionFilter,
-        aspectRatioStrategy: aspectRatioStrategy);
-  }
+  static AspectRatioStrategy
+  _ratio_16_9FallbackAutoStrategyAspectRatioStrategy() =>
+      AspectRatioStrategy.ratio_16_9FallbackAutoStrategy;
 
-  static FallbackStrategy _createAttachedFallbackStrategy(
-      {required VideoQuality quality,
-      required VideoResolutionFallbackRule fallbackRule}) {
-    return FallbackStrategy(quality: quality, fallbackRule: fallbackRule);
-  }
+  static AspectRatioStrategy
+  _ratio_4_3FallbackAutoStrategyAspectRatioStrategy() =>
+      AspectRatioStrategy.ratio_4_3FallbackAutoStrategy;
 
-  static QualitySelector _createAttachedQualitySelector(
-      {required VideoQuality videoQuality,
-      required FallbackStrategy fallbackStrategy}) {
-    return QualitySelector.from(
-        quality: VideoQualityData(quality: videoQuality),
-        fallbackStrategy: fallbackStrategy);
-  }
+  static CaptureRequestKey _controlAELockCaptureRequest() =>
+      CaptureRequest.controlAELock;
 
-  static Future<void> _requestCameraPermissions(bool enableAudio) async {
-    await SystemServices.requestCameraPermissions(enableAudio);
-  }
+  static CameraCharacteristicsKey
+  _infoSupportedHardwareLevelCameraCharacteristics() =>
+      CameraCharacteristics.infoSupportedHardwareLevel;
 
-  static void _startListeningForDeviceOrientationChange(
-      bool cameraIsFrontFacing, int sensorOrientation) {
-    DeviceOrientationManager.startListeningForDeviceOrientationChange(
-        cameraIsFrontFacing, sensorOrientation);
-  }
-
-  static Future<int> _setPreviewSurfaceProvider(Preview preview) async {
-    return preview.setSurfaceProvider();
-  }
-
-  static Future<int> _getDefaultDisplayRotation() async {
-    return DeviceOrientationManager.getDefaultDisplayRotation();
-  }
-
-  static Camera2CameraControl _getCamera2CameraControl(
-      CameraControl cameraControl) {
-    return Camera2CameraControl(cameraControl: cameraControl);
-  }
-
-  static CaptureRequestOptions _createAttachedCaptureRequestOptions(
-      List<(CaptureRequestKeySupportedType, Object?)> options) {
-    return CaptureRequestOptions(requestedOptions: options);
-  }
-
-  static MeteringPoint _createAttachedMeteringPoint(
-      double x, double y, double? size, CameraInfo cameraInfo) {
-    return MeteringPoint(x: x, y: y, size: size, cameraInfo: cameraInfo);
-  }
-
-  static FocusMeteringAction _createAttachedFocusMeteringAction(
-      List<(MeteringPoint, int?)> meteringPointInfos, bool? disableAutoCancel) {
-    return FocusMeteringAction(
-        meteringPointInfos: meteringPointInfos,
-        disableAutoCancel: disableAutoCancel);
-  }
-
-  static AspectRatioStrategy _createAttachedAspectRatioStrategy(
-      int preferredAspectRatio, int fallbackRule) {
-    return AspectRatioStrategy(
-        preferredAspectRatio: preferredAspectRatio, fallbackRule: fallbackRule);
-  }
-
-  static ResolutionFilter _createAttachedResolutionFilterWithOnePreferredSize(
-      Size preferredSize) {
-    return ResolutionFilter.onePreferredSize(
-        preferredResolution: preferredSize);
-  }
-
-  static Future<Camera2CameraInfo> _getCamera2CameraInfo(
-      CameraInfo cameraInfo) async {
-    return Camera2CameraInfo.from(cameraInfo);
-  }
-
-  static Future<DeviceOrientation> _getUiOrientation() async {
-    return DeviceOrientationManager.getUiOrientation();
-  }
-
-  static Future<int> _getSensorOrientation(
-      Camera2CameraInfo camera2CameraInfo) async {
-    return camera2CameraInfo.getSensorOrientation();
-  }
-
-  static Future<bool> _previewSurfaceProducerHandlesCropAndRotation(
-      Preview preview) async {
-    return preview.surfaceProducerHandlesCropAndRotation();
-  }
+  static CameraCharacteristicsKey _sensorOrientationCameraCharacteristics() =>
+      CameraCharacteristics.sensorOrientation;
 }
